Adding MCPhylo

Parsing Functionality

Tree Manipulation

Return Functions

Adding MCPhylo

julia> using Pkg

julia> Pkg.add("MCPhylo")
ERROR: The following package names could not be resolved:
 * MCPhylo (not found in project, manifest or registry)

Parsing Functionality

Main.MCPhylo.ParseNewick — Function

ParseNewick(filename::String)

This is the main function, which parses a file, containing Newick strings.

Main.MCPhylo.ParseCSV — Function

ParseCSV(filename::String)

This function parses a CSV file which stores the input for the MCMC compuation. The file should follow the conventions used for MrBayes.

Main.MCPhylo.ParseNexus — Function

ParseNexus(filename::String)

This function parses a NEXUS file which stores the input for the MCMC compuation. The file should follow the conventions used for MrBayes.

Tree Manipulation

Main.MCPhylo.move! — Function

move!(node1::Node, node2::Node, proportion::Float64)

Change the incomming length of node1 and node2 while keeping there combined length constant.

Main.MCPhylo.merge_trees! — Function

merge_trees!(ref_tree::T, tree::T)::Tuple{T, Vector{T}} where T<:AbstractNode

Merge two compatible trees, i.e. inserts all cluster of the first tree, which aren't already in the second tree, into the secon tree

Main.MCPhylo.ladderize_tree! — Function

ladderize_tree!(root::T, ascending::Bool=true)::Nothing where T<:AbstractNode)

This function ladderizes a tree inplace, i.e. sorts the nodes on all levels by the count of their descendants

Main.MCPhylo.prune_tree! — Function

prune_tree!(root::T, node_names::Vector{String})::Nothing where T<:AbstractNode

In-place version of prune_tree

prune_tree!(root::T, node_names::Vector{T})::Nothing where T<:AbstractNode

In-place version of prune_tree

Main.MCPhylo.prune_tree — Function

prune_tree(root::T, node_names::Vector{String})::T where T<:AbstractNode

This function returns a copy of a tree with specific nodes including their descendants removed

Main.MCPhylo.add_child! — Function

add_child!(mother_node::Node, child::Node, child_position::Union{Int64, Missing}=missing)

This function adds a child to the mother node. The arity of the mother node is increased by 1 and the root status of the child is set to False.

Main.MCPhylo.remove_child! — Function

remove_child!(mother_node::Node, left::Bool)::Node

This function removes a child from the list of nodes which are daughters of this node. The removed node is returned.

remove_child!(mother_node::Node, child::Node)::Node

This function removes a child from the list of nodes which are daughters of this node. The removed node is returned.

Main.MCPhylo.create_tree_from_leaves — Function

create_tree_from_leaves(leaf_nodes::Vector{T})::Node

This function creates a random binary tree from a list of leaf nodes. The root node as access point for the tree is returned.

Main.MCPhylo.create_tree_from_leaves_bin — Function

create_tree_from_leaves(leaf_nodes::Vector{T})::Node

This function creates a random binary tree from a list of leaf nodes. The root node as access point for the tree is returned.

Main.MCPhylo.insert_node! — Function

insert_node!(mother::Node, children::Vector{T})::T where T<:AbstractNode

This function inserts a node into a tree after a mother node and gains a subset of the mother's children as its children. Returns the inserted node.

Main.MCPhylo.number_nodes! — Function

number_nodes!(root::T)::Nothing  where T<:GeneralNode

This function assigns a unique, sequential number to each node.

Main.MCPhylo.perform_spr — Function

perform_spr(root::Node)

performs SPR on binary tree

Main.MCPhylo.risky_SPR — Function

    risky_SPR(root::Node)::AbstractNode
Performs SPR on tree in place; takes reference to root of tree, boolean value necessary to determine if tree should be treated as binary or not
Returns copy of root of altered tree. Does not check for correct formatting of tree.

Main.MCPhylo.risky_SPR! — Function

    risky_SPR!(root::Node)::AbstractNode
Performs SPR on tree in place; takes reference to root of tree, boolean value necessary to determine if tree should be treated as binary or not
Returns reference to root of altered tree. Does not check for correct formatting of tree.

Return Functions

Main.MCPhylo.newick — Function

newick(node::Node)::String

Creates a newick represnetation of the tree.

newick(root::T, newickstring::AbstractString) where T<:GeneralNode

Do the newick recursion. It is meant as the internal iterator function.

Main.MCPhylo.tree_height — Function

tree_height(root::T)::Float64  where T<:GeneralNode

This function calculates the tree height.

Main.MCPhylo.level_order — Function

level_order(node::T)::Array{T} where T<:GeneralNode

This function does level order traversal. Only the root node needs to be supplied.

Main.MCPhylo.tree_length — Function

tree_length(root::T)::Float64  where T<:GeneralNode

This function calculates the tree length.

tree_length(root::T, tl::Float64)::Float64 where T<:GeneralNode

This function does the internal tree length recursion

Main.MCPhylo.get_leaves — Function

post_order(root::T, traversal::Vector{T})::Vector{T} where T<:GeneralNode

This function performs a post order traversal through the tree. It is assumed that root is the root of the tree. Thus, if root is not the root, the subtree defined by the root root is used for the post order traversal.

post_order(root::T)::Vector{T} where T<:GeneralNode

This function does post order traversal. Only the root node needs to be supplied.

Main.MCPhylo.get_bipartitions — Function

get_bipartitions(tree::T)::Vector{Tuple} where T <:GeneralNode

Get a vector of all bipartions of tree. The resulting vector contains Tuples of sets representing the bipartions.

Main.MCPhylo.get_sister — Function

get_sister(node::T)::T  where T<:GeneralNode

This function gets the sister of node. It does so by looking for the respective binary representation of the sister.

Main.MCPhylo.get_branchlength_vector — Function

get_branchlength_vector(root::N)::Vector{T}  where {N <:GeneralNode, T<:Real}

Get the vector of branch lengths of the tree.

get_branchlength_vector(t::TreeStochastic)

Get the vector of branch lengths of the tree.

get_branchlength_vector(root::N, out_vec::Vector{T}) where {N<:GeneralNode, T<:Real}

Do post order traversal to retrieve a vector of branch lengths.

get_branchlength_vector(post_order::Vector{Node})::Vector{Float64}

Return a vector of branch lengths.

Main.MCPhylo.get_mother — Function

function get_mother(node::T)::T  where T<:GeneralNode

This function gets the mother of node. It does so by looking for the respective binary representation of the mother node.

Main.MCPhylo.get_sum_seperate_length! — Function

get_sum_seperate_length!(root::Node)::Vector{Float64}

This function gets the sum of the branch lengths of the internal branches and the branches leading to the leave nodes.

get_sum_seperate_length!(root::Node)::Vector{Float64}

This function gets the sum of the branch lengths of the internal branches and the branches leading to the leave nodes.