Tree Functionality

Functions found in all files of Tree folder, ordered according to file of origin.

Converter.jl

Main.MCPhylo.from_df — Function

from_df(df::DataFrame)::Node

This function takes a DataFrame and turns it into a tree. It assumes a rooted binary tree is stored in the matrix. No checks are performed.

Main.MCPhylo.newick — Method

newick(node::Node)::String

Creates a newick represnetation of the tree.

Main.MCPhylo.to_covariance — Function

to_covariance(tree::N, blv::Array{T})::Array{T,2} where {N<:GeneralNode,T<: Real}

Calcualte the variance-covariance matrix from tree. An entry (i,j) of the matrix is defined as the length of the path connecting the latest common ancestor of i and j with the root of the tree.

Main.MCPhylo.to_covariance_ultra — Function

to_covariance_ultra(tree::Node)::Array{T,2} where T<: Real

Get the covariance matrix of the ultrametric version of tree with height 1.

Main.MCPhylo.to_df — Function

to_df(root::Node)::DataFrame

This function returns a matrix representation of the tree structure. The matirx is returned as a DataFrame so that the names of the columns are the names of the tips in the tree. The entry df[i,j] is the length of the edge connecting node i with node j.

Main.MCPhylo.to_distance_matrix — Function

to_distance_matrix(tree::T)::Array{Float64,2} where T <:GeneralNode

Calculate the distance matrix over the set of leaves.

Node_Type.jl

Tree_Basics.jl

This file is a bit larger, and so will be broken into (hopefully coherent) subgroups, listed here:

Tree Editing

Return Functions

Metavariable Functions

Variable Functions

Tree Editing

These functions edit the makeup of an existing tree.

Main.MCPhylo.add_child! — Function

add_child!(mother_node::Node, child::Node, child_position::Union{Int64, Missing}=missing)

This function adds a child to the mother node. The arity of the mother node is increased by 1 and the root status of the child is set to False.

Main.MCPhylo.delete_node! — Function

delete_node!(node::Node)::Nothing

This functions deletes node from a tree and assigns all its children to its mother node.

Main.MCPhylo.insert_node! — Function

insert_node!(mother::Node, children::Vector{T})::T where T<:AbstractNode

This function inserts a node into a tree after a mother node and gains a subset of the mother's children as its children. Returns the inserted node.

Main.MCPhylo.remove_child! — Function

remove_child!(mother_node::Node, left::Bool)::Node

This function removes a child from the list of nodes which are daughters of this node. The removed node is returned. An input of "True" removes the left child, while "False" removes the right child.

remove_child!(mother_node::Node, child::Node)::Node

This function removes a child from the list of nodes which are daughters of this node. The removed node is returned.

Main.MCPhylo.create_tree_from_leaves — Function

create_tree_from_leaves(leaf_nodes::Vector{T})::Node

This function creates a random binary tree from a list of leaf nodes. The root node as access point for the tree is returned.

Return Functions

These functions return the values of a given variable of a tree.

Main.MCPhylo.get_mother — Function

function get_mother(node::T)::T  where T<:GeneralNode

This function gets the mother of node. It does so by looking for the respective binary representation of the mother node.

Main.MCPhylo.get_sister — Function

get_sister(node::T)::T  where T<:GeneralNode

This function gets the sister of node. It does so by looking for the respective binary representation of the sister.

Main.MCPhylo.node_height — Function

node_height(root::T, mv::Float64)::Float64  where T<:GeneralNode

Calculate the height of a node.

Main.MCPhylo.random_node — Function

function random_node(root::T)::T  where T<:GeneralNode

This function returns a random node from the tree.

Metavariable Functions

These functions calculate metavariables for a given tree.

Main.MCPhylo.tree_length — Method

tree_length(root::T)::Float64  where T<:GeneralNode

This function calculates the tree length.

Main.MCPhylo.tree_height — Function

tree_height(root::T)::Float64  where T<:GeneralNode

This function calculates the tree height.

Main.MCPhylo.get_branchlength_vector — Function

get_branchlength_vector(root::N)::Vector{T}  where {N <:GeneralNode, T<:Real}

Get the vector of branch lengths of the tree.

get_branchlength_vector(t::TreeStochastic)

Get the vector of branch lengths of the tree.

get_branchlength_vector(root::N, out_vec::Vector{T}) where {N<:GeneralNode, T<:Real}

Do post order traversal to retrieve a vector of branch lengths.

get_branchlength_vector(post_order::Vector{Node})::Vector{Float64}

Return a vector of branch lengths.

Main.MCPhylo.get_sum_seperate_length! — Function

get_sum_seperate_length!(root::T)::Vector{Float64}  where T<:GeneralNode

This function gets the sum of the branch lengths of the internal branches and the branches leading to the leave nodes.

get_sum_seperate_length!(post_order::Vector{T})::Vector{Float64}  where T<:GeneralNode

This function gets the sum of the branch lengths of the internal branches and the branches leading to the leave nodes.

Main.MCPhylo.check_binary — Function

check_binary(root::Node)::Bool

checks to see if given tree is binary; returns true if properly formatted and false otherwise

Main.MCPhylo.path_length — Function

path_length(ancestor::T, descendant::T)::Float64  where T<:GeneralNode

Note: The function assumes there is an ancestral relationship between the two nodes.

This function calculates the length of the path separating the ancestor from the offspring node. The function follows the path specified through the binary description of the node.

Variable Functions

These Functions calculate/edit variables for a given tree.

Main.MCPhylo.force_ultrametric! — Function

force_ultrametric!(root::T) where T<:GeneralNode

Force an ultrametric version of the tree.

Main.MCPhylo.number_nodes! — Function

number_nodes!(root::T)::Nothing  where T<:GeneralNode

This function assigns a unique, sequential number to each node.

Main.MCPhylo.set_branchlength_vector! — Function

set_branchlength_vector!(t::TreeStochastic, blenvec::Array{T}) where T <: Real

Get the vector of branch lengths of the tree.

set_branchlength_vector!(t::TreeStochastic, blenvec::ArrayStochastic)

Get the vector of branch lengths of the tree.

set_branchlength_vector!(root::N, blenvec::Array{T}) where {N<:GeneralNode, T<:Real}

This function sets the branch lengths of a tree to the values specified in blenvec.

Main.MCPhylo.set_binary! — Function

set_binary!(root::Node)

Assign a binary representation to each node, which specifies the path from the root to this node via the binary representation of the node. A left turn is a 1 in binary and a right turn a 0.

Tree_Clustering.jl

Main.MCPhylo.neighbor_joining — Function

neighbor_joining(dm::Array{Float64,2}, Array{String,1})

This function returns a phylogenetic tree by using neighbor-joining based on a given distance matrix and an array of leaf names. Returns a node of the resulting tree, from which it can be traversed.

neighbor_joining(dm::Array{Float64,2})

This function returns a phylogenetic tree by using neighbor-joining based on a given distance matrix. Creates an array of nodes to be used as leaves. Returns a node of the resulting tree, from which it can be traversed.

Main.MCPhylo.upgma — Function

upgma(dm::Array{Float64,2}, Array{String,1})

This function returns a phylogenetic tree by using UPGMA based on a given distance matrix and an array of leaf names. Returns a node of the resulting tree, from which it can be traversed.

upgma(dm::Array{Float64,2})

This function returns a phylogenetic tree by using UPGMA based on a given distance matrix. Creates an array of nodes to be used as leaves. Returns a node of the resulting tree, from which it can be traversed.

Tree_Consensus.jl

Main.MCPhylo.find_common_clusters — Method

find_common_clusters(ref_tree, tree:T)
    ::Dict{Int64, Tuple{Bool, Union{Float64, Missing}}}

Use Day's algorithm to create a dictionary, that tells us for each node of the second input tree, if its corresponding cluster is a common cluster of the trees

Main.MCPhylo.majority_consensus_tree — Method

majority_consensus_tree(trees::Vector{T}, percentage::Float64=0.5)::T where T<:AbstractNode

Construct the majority rule consensus tree from a set of trees. By default includes cluster that occur in over 50% of the trees.

Main.MCPhylo.max_leaf_rank — Method

max_leaf_rank(leaf_ranks::Dict{String, Int64}, node::T)
    ::Int64 where T<:AbstractNode

Recursive helper function to find the highest ranked leaf descendant of a node

Main.MCPhylo.merge_trees! — Method

merge_trees!(ref_tree::T, tree::T)::Tuple{T, Vector{T}} where T<:AbstractNode

Merge two compatible trees, i.e. inserts all cluster of the first tree, which aren't already in the second tree, into the secon tree

Main.MCPhylo.min_leaf_rank — Method

min_leaf_rank(leaf_ranks::Dict{String, Int64}, node::T)
    ::Int64 where T <: AbstractNode

Recursive helper function to find the lowest ranked leaf descendant of a node

Main.MCPhylo.one_way_compatible — Method

one_way_compatible(ref_tree::T, tree::T)::T where T<:AbstractNode

Takes two trees and returns a copy of the first one, where all the clusters that are not compatible with the second tree are removed.

Main.MCPhylo.order_tree! — Method

order_tree!(root::T, cluster_start_indeces::Dict{T, Int64}, leaves=Vector{T}())
    ::Vector{T} where T<:AbstractNode

Helper function to order a tree based on cluster indeces and return the leaves of the ordered tree

Main.MCPhylo.x_left — Method

x_left(node::T)::Tuple{T,Vector{T}} where T<:AbstractNode

Helper function to find ancestor of a leaf that has said leaf as leftmost descendant. Also returns the path from the leaf to the mother of that node.

Main.MCPhylo.x_right — Method

x_right(node::T)::Tuple{T,Vector{T}} where T<:AbstractNode

Helper function to find ancestor of a leaf that has said leaf as rightmost descendant. Also returns the path from the leaf to the mother of that node.

Tree_Distance.jl

Main.MCPhylo.BHV_bounds — Method

BHV_bounds(tree1::T, tree2::T)::Tuple{Float64, Float64} where T <:GeneralNode

This function calculates the lower and upper bounds of the geodesic in the Billera-Holmes-Vogtman space.

Main.MCPhylo.RF — Method

RF(tree1::T, tree2::T)::Int64 where T <:GeneralNode

Calculate the Robinson-Foulds distance between the two trees. In its current form the function assumes the trees have identical leave sets.

Main.MCPhylo.get_bipartitions — Method

get_bipartitions(tree::T)::Vector{Tuple} where T <:GeneralNode

Get a vector of all bipartions of tree. The resulting vector contains Tuples of sets representing the bipartions.

Tree_Ladderizing.jl

Main.MCPhylo.ladderize_tree! — Method

ladderize_tree!(root::T, ascending::Bool=true)::Nothing where T<:AbstractNode)

This function ladderizes a tree inplace, i.e. sorts the nodes on all levels by the count of their descendants

Main.MCPhylo.ladderize_tree — Method

ladderize_tree(root::T, ascending::Bool=true)::T where T<:AbstractNode)

This function returns a ladderized copy of a tree, i.e. a copy with all the nodes on all levels sorted by the count of their descendants

Tree_Legacy.jl

Main.MCPhylo.get_branchlength_vector — Method

get_branchlength_vector(post_order::Vector{Node})::Vector{Float64}

Return a vector of branch lengths.

Tree_moves.jl

Main.MCPhylo.NNI! — Function

NNI(root::T, target::T, lor::Bool)::Int64   where T<:GeneralNode

This function does a nearest neighbour interchange (NNI) move on the tree specified by root. The parameter target specifies the node which performs the interchange move using the left or right child of the target node. If the left child should be used lor=true. The function returns 1 if the move was successful and 0 else.

NNI!(root::T, target::Int64)::Int64  where T<:GeneralNode

This function does a nearest neighbour interchange (NNI) move on the tree specified by root. The target is identified by the number of the target node. The function returns 1 if the move was successful and 0 else.

NNI!(root::T)::Int64  where T<:GeneralNode

This function does a nearest neighbour interchange (NNI) move on the tree specified by root. The target is identified by the number of the target node. The function returns 1 if the move was successful and 0 else.

Main.MCPhylo.NNI — Function

NNI(root::T)::T  where T<:GeneralNode

This function does a nearest neighbour interchange (NNI) move on the tree specified by root and returns a mutated copy while leaving the original tree intact.

Main.MCPhylo.change_edge_length! — Function

change_edge_length!(root::T) where T <:GeneralNode

Pick a random node and increase or decrease its length randomly.

Main.MCPhylo.move! — Function

move!(node1::Node, node2::Node, proportion::Float64)

Change the incoming length of node1 and node2 while keeping their combined length constant.

Main.MCPhylo.perform_spr — Function

perform_spr(root::Node)

performs SPR on binary tree

Main.MCPhylo.randomize! — Function

randomize!(root::Node, num::Int64=100)::nothing

This function randomizes the tree topology by performing a number of nearest neighbour interchange (NNI) moves. The number of NNI moves is specified in the parameter num.

Main.MCPhylo.risky_SPR! — Function

    risky_SPR!(root::Node)::AbstractNode
Performs SPR on tree in place. Takes reference to root of tree
Returns reference to root of altered tree. Does not check for correct formatting of tree.

Main.MCPhylo.risky_SPR — Function

    risky_SPR(root::Node)::AbstractNode
Performs SPR on tree in place. Takes reference to root of tree
Returns copy of root of altered tree. Does not check for correct formatting of tree.

Main.MCPhylo.slide — Function

slide(root::T)::T where T<:GeneralNode

This functin performs a slide move on an intermediate node. The node is moved upwards or downwards on the path specified by its mother and one of its daughters. The new tree is returned.

Main.MCPhylo.slide! — Function

slide!(root::T) where T<:GeneralNode

This functin performs a slide move on an intermediate node. The node is moved upwards or downwards on the path specified by its mother and one of its daughters.

Main.MCPhylo.swing! — Function

swing!(root::T) where T<:GeneralNode

This function performs a swing node. A random non-leave node is selected and moved along the path specified by its two children.

Main.MCPhylo.swing — Function

swing(root::T)::T where T<:GeneralNode

This function performs a swing node. A random non-leave node is selected and moved along the path specified by its two children. The new tree is returned.

Tree_Pruning.jl

Main.MCPhylo.prune_tree! — Function

prune_tree!(root::T, node_names::Vector{String})::Nothing where T<:AbstractNode

In-place version of prune_tree

prune_tree!(root::T, node_names::Vector{T})::Nothing where T<:AbstractNode

In-place version of prune_tree

Main.MCPhylo.prune_tree — Function

prune_tree(root::T, node_names::Vector{String})::T where T<:AbstractNode

This function returns a copy of a tree with specific nodes including their descendants removed

Tree_Search.jl

Main.MCPhylo.find_binary — Function

find_binary(root::T, bin::String)::T where T<:GeneralNode

Find a node by its binary representation. The function assumes that the node is present in the tree.

Do not use this function if you are unsure wheter the node is in the tree at all.

Main.MCPhylo.find_num — Function

find_num(root::T, num::Int64)  where T<:GeneralNode

Find a node by its number. The function assumes that the node is present in the tree.

Do not use this function if you are unsure wheter the node is in the tree at all.

find_num(root::T, num::Int64, rn::Vector{T})::Bool  where T<:GeneralNode

Do a post order traversal to find the node corresponding to the num.

Main.MCPhylo.find_root — Function

find_root(node::Node)::Node

Takes a node in a tree and finds the root of that tree

Tree_Traversal.jl

Main.MCPhylo.get_leaves — Function

post_order(root::T, traversal::Vector{T})::Vector{T} where T<:GeneralNode

This function performs a post order traversal through the tree. It is assumed that root is the root of the tree. Thus, if root is not the root, the subtree defined by the root root is used for the post order traversal.

post_order(root::T)::Vector{T} where T<:GeneralNode

This function does post order traversal. Only the root node needs to be supplied.

Main.MCPhylo.level_order — Function

level_order(node::T)::Array{T} where T<:GeneralNode

This function does level order traversal. Only the root node needs to be supplied.

Main.MCPhylo.level_traverse — Function

level_traverse(node::T, level::Int64, stack::Array{T})::Bool where T <:GeneralNode

This function traverses a level of the tree specified through node. The level is specified via the level argument and the nodes visited are stored in the stack. This function is intended as the internal worker for the level_order function.

Main.MCPhylo.post_order — Function

post_order(root::T, traversal::Vector{T})::Vector{T} where T<:GeneralNode

This function performs a post order traversal through the tree. It is assumed that root is the root of the tree. Thus, if root is not the root, the subtree defined by the root root is used for the post order traversal.

post_order(root::T)::Vector{T} where T<:GeneralNode

This function does post order traversal. Only the root node needs to be supplied.

Main.MCPhylo.pre_order — Function

pre_order(root::T, traversal::Vector{T})::Vector{T} where T<:GeneralNode

This function performs a pre order traversal through the tree. It is assumed that root is the root of the tree. Thus, if root is not the root, the subtree defined by the root root is used for the pre order traversal.

pre_order(root::T)::Vector{T} where T<:GeneralNode

This function does pre order traversal. Only the root node needs to be supplied.