List of all functions (minus private ones) in the alphabetical order, capital letters first

Main.MCPhylo.RWM — Method

RWM(params::ElementOrVector{Symbol}, moves::Array{Symbol}; args...)

Construct the RWM sampler for Trees. If you set moves to :all it will use all eligible moves to change the tree. These are currently: NNI, Slide, Swing, :EdgeLength

Main.MCPhylo.prune_tree! — Method

prune_tree!(root::T, node_names::Vector{String})::Nothing where T<:AbstractNode

In-place version of prune_tree

Main.MCPhylo.prune_tree! — Method

prune_tree!(root::T, node_names::Vector{T})::Nothing where T<:AbstractNode

In-place version of prune_tree

Main.MCPhylo.prune_tree — Method

prune_tree(root::T, node_names::Vector{String})::T where T<:AbstractNode

This function returns a copy of a tree with specific nodes including their descendants removed