output

Chains

Main.MCPhylo.Chains — Type

Chains(iters::Integer, params::Integer;
start::Integer=1, thin::Integer=1, chains::Integer=1,
names::Vector{T}=AbstractString[]) where {T<:AbstractString}

Chains(value::Array{T, 3},
start::Integer=1, thin::Integer=1,
names::Vector{W}=AbstractString[], chains::Vector{V}=Int[], moves::Vector{V}=Int[0])
where {T<:Real, U<:AbstractString, V<:Integer, W <: AbstractString}

Chains(value::Array{T, 3},
value2::Array{U,3};
start::Integer=1, thin::Integer=1,
names::Vector{W}=AbstractString[], chains::Vector{V}=Int[], moves::Vector{V}=Int[0])
where {T<:Real, U<:AbstractString, V<:Integer, W <: AbstractString}

Chains(value::Matrix{T};
start::Integer=1, thin::Integer=1,
names::Vector{U}=AbstractString[], chains::Integer=1)
where {T<:Real, U<:AbstractString}

Chains(value::Vector{T};
start::Integer=1, thin::Integer=1,
names::U="Param1", chains::Integer=1) where {T<:Real, U <: AbstractString}

Construct a Chains object that stores MCMC sampler output.

Returns an object of type Chains.

iters: total number of iterations in each sampler run, of which length(start:thin:iters) outputted iterations will be stored in the object.

params: number of parameters to store.

value: array whose first, second (optional), and third (optional) dimensions index outputted iterations, parameter elements, and runs of an MCMC sampler, respectively.

start: number of the first iteration to be stored.

thin: number of steps between consecutive iterations to be stored.

chains: number of simulation runs for which to store output, or indices to the runs (default: 1, 2, …).

names: names to assign to the parameter elements (default: "Param1", "Param2", …).

Base.getindex — Function

Base.getindex(c::Chains, window, names, chains)

Subset MCMC sampler output. The syntax c[i, j, k] is converted to getindex(c, i, j, k).

Subsetted sampler output stored in the same type of object as that supplied in the call.

c : sampler output to subset.

window : indices of the form start:stop or start:thin:stop can be used to subset iterations, where start and stop define a range for the subset and thin will apply additional thinning to existing sampler output.

names : indices for subsetting of parameters that can be specified as strings, integers, or booleans identifying parameters to be kept. ModelChains may additionally be indexed by model node symbols.

chains : indices for chains can be integers or booleans.

A value of : can be specified for any of the dimensions to indicate no subsetting.

Base.setindex! — Function

Base.setindex!(c::AbstractChains, value, iters, names, chains)

Store MCMC sampler output at a given index. The syntax c[i, j, k] = value is converted to setindex!(c, value, i, j, k).

Returns an object of the same type as c with the sampler output stored in the specified indices.

c : object within which to store sampler output.

value : sampler output.

iters : iterations can be indexed as a start:stop or start:thin:stop range, a single numeric index, or a vector of indices; and are taken to be relative to the index range store in the c.range field.

names : indices for subsetting of parameters can be specified as strings, integers, or booleans.

chains : indices for chains can be integers or booleans.

A value of : can be specified for any of the dimensions to indicate no subsetting.

Base.cat — Function

Base.cat(c1::AbstractChains, args::AbstractChains...; dims::Integer)

Concatenate input MCMC chains along a specified dimension. For dimensions other than the specified one, all input chains must have the same sizes, which will also be the sizes of the output chain. The size of the output chain along the specified dimension will be the sum of the sizes of the input chains in that dimension. vcat concatenates vertically along dimension 1, and has the alternative syntax [chain1; chain2; ...]. hcat concatenates horizontally along dimension 2, and has the alternative syntax [chain1 chain2 ...].

Returns a Chains object containing the concatenated input.

dim : dimension (1, 2, or 3) along which to concatenate the input chains.

c1, args... : Chains to concatenate.

Base.keys — Method

Base.keys(c::AbstractChains)

Returns names of parameter elements.

c : Chain to return names of.

Base.show — Method

Base.show(io::IO, c::AbstractChains)

Prints header and values of Chain.

io : IO stream on which to print.

AbstractChains : Chain to print.

Base.size — Function

Base.size(c::AbstractChains)

Returns Tuple containing last iteration of MCMC sampler output and dimensions of Chain dimensions of c.value.

c : Chain object of interest.

Base.size(c::AbstractChains, ind)

Returns last iteration of MCMC sampler output, or dimension derived from C, according to value of ind.

c : Chain object of interest.

ind : index of tuple to return; 1 returns last iteration of MCMC sampler output, 2 and 3 return dimensions of c.value.

File I/O

Base.read — Method

Base.read(name::AbstractString, ::Type{T}) where {T<:AbstractChains}

Read a chain from an external file.

Returns an AbstractChains subtype read from an external file.

name : file to read or write. Recommended convention is for the file name to be specified with a .jls extension.

T : chain type to read.

c : chain to write.

Base.write — Method

Base.write(name::AbstractString, c::AbstractChains)

Write a chain to an external file.

Returns a written external file containing a subtype.

name : file to read or write. Recommended convention is for the file name to be specified with a .jls extension.

T : chain type to read.

c : chain to write.

Main.MCPhylo.readcoda — Method

readcoda(output::AbstractString, index::AbstractString)

Read MCMC sampler output generated in the CODA format by OpenBUGS. The function only retains those sampler iterations at which all model parameters were monitored.

Returns a Chains object containing the read sampler output.

output : text file containing the iteration numbers and sampled values for the model parameters.

index : text file containing the names of the parameters, followed by the first and last rows in which their output can be found in the output file.

Discrete Diagnostics

Main.MCPhylo.discretediag — Method

discretediag(c::AbstractChains; frac::Real=0.3,
              method::Symbol=:weiss, nsim::Int=1000)

Compute the convergence diagnostic for a discrete variable. Several options are available by choosing method to be one of :hangartner, :weiss, :DARBOOT, MCBOOT, :billinsgley, :billingsleyBOOT. The first four are based off of Pearson’s chi-square test of homogeneity. The diagnostic tests whether the proportion of the categories of the discrete variable are similar in each chain. The last two methods test whether the transition probabilities between each category are similar between each chain. Along with a between chain assessment of convergence, a within-chain assessment is carried out by comparing a specified fraction (frac), or window, of the beginning of a chain to the specified fraction of the end of the chain. For within-chain assessment, users should ensure that there is sufficient separation between the windows to assume that their samples are independent. A non-significant test p-value indicates convergence. Significant p-values indicate non-convergence and the possible need to discard initial samples as a burn-in sequence or to simulate additional samples.

A ChainSummary type object with parameters contained in the rows of the value field. The first three columns correspond to the test statistic, degrees of freedom, and p-value of the between-chain assessment. The next columns are the test statistic, degrees of freedom, and p-value for each chain of the within-chain assessment.

c : sampler output on which to perform calculations.

frac : proportion of iterations to include in the first window.

method : Specify which method to use. One of :hangartner, :weiss, :DARBOOT, MCBOOT, :billinsgley, :billingsleyBOOT`.

nsim : For the bootstrap methods (:DARBOOT, :MCBOOT, and :billingsleyBOOT) the number of bootstrap simulations.

Gelman Diagnostic

Main.MCPhylo.gelmandiag — Method

gelmandiag(c::AbstractChains; alpha::Real=0.05, mpsrf::Bool=false,
            transform::Bool=false)

Compute the convergence diagnostics of Gelman, Rubin, and Brooks for MCMC sampler output. The diagnostics are designed to asses convergence of posterior means estimated with multiple autocorrelated samples (chains). They does so by comparing the between and within-chain variances with metrics called potential scale reduction factors (PSRF). Both univariate and multivariate factors are available to assess the convergence of parameters individually and jointly. Scale factors close to one are indicative of convergence. As a rule of thumb, convergence is concluded if the 0.975 quantile of an estimated factor is less than 1.2. Multiple chains are required for calculations. It is recommended that at least three chains be generated, each with different starting values chosen to be diffuse with respect to the anticipated posterior distribution. Use of multiple chains in the diagnostic provides for more robust assessment of convergence than is possible with single chain diagnostics.

Returns a ChainSummary type object of the form:

struct ChainSummary
  value::Array{Float64, 3}
  rownames::Vector{AbstractString}
  colnames::Vector{AbstractString}
  header::AbstractString
end

with parameters contained in the rows of the value field, and scale reduction factors and upper-limit quantiles in the first and second columns.

c : sampler output on which to perform calculations.

alpha : quantile (1 - alpha / 2) at which to estimate the upper limits of scale reduction factors.

mpsrf : whether to compute the multivariate potential scale reduction factor. This factor will not be calculable if any one of the parameters in the output is a linear combination of others.

transform : whether to apply log or logit transformations, as appropriate, to parameters in the chain to potentially produce output that is more normally distributed, an assumption of the PSRF formulations.

Geweke Diagnostic

Main.MCPhylo.gewekediag — Function

gewekediag(x::Vector{T}; first::Real=0.1, last::Real=0.5,
            etype=:imse, args...) where {T<:Real}

gewekediag(c::AbstractChains; first::Real=0.1, last::Real=0.5,
            etype=:imse, args...)

Compute the convergence diagnostic of Geweke [37] for MCMC sampler output. The diagnostic is designed to asses convergence of posterior means estimated with autocorrelated samples. It computes a normal-based test statistic comparing the sample means in two windows containing proportions of the first and last iterations. Users should ensure that there is sufficient separation between the two windows to assume that their samples are independent. A non-significant test p-value indicates convergence. Significant p-values indicate non-convergence and the possible need to discard initial samples as a burn-in sequence or to simulate additional samples.

Returns a ChainSummary type object with parameters contained in the rows of the value field, and test Z-scores and p-values in the first and second columns. Results are chain-specific.

c : sampler output on which to perform calculations.

first : proportion of iterations to include in the first window.

last : proportion of iterations to include in the last window.

etype : method for computing Monte Carlo standard errors. See mcse() for options.

args... : additional arguments to be passed to the etype method.

x : vector on which to perform calculations.

Heidel Diagnostic

Main.MCPhylo.heideldiag — Function

heideldiag(x::Vector{T}; alpha::Real=0.05, eps::Real=0.1,
            etype=:imse, start::Integer=1, args...) where {T<:Real}

heideldiag(c::AbstractChains; alpha::Real=0.05, eps::Real=0.1,
            etype=:imse, args...)

Compute the convergence diagnostic of Heidelberger and Welch for MCMC sampler output. The diagnostic is designed to assess convergence of posterior means estimated with autocorrelated samples and to determine whether a target degree of accuracy is achieved. A stationarity test is performed for convergence assessment by iteratively discarding 10% of the initial samples until the test p-value is non-significant and stationarity is concluded or until 50% have been discarded and stationarity is rejected, whichever occurs first. Then, a halfwidth test is performed by calculating the relative halfwidth of a posterior mean estimation interval as $z_{1 - \alpha / 2} \hat{s} / |\bar{\theta}|$; where $z$ is a standard normal quantile, $\hat{s}$ is the Monte Carlo standard error, and $\bar{\theta}$ is the estimated posterior mean. If the relative halfwidth is greater than a target ratio, the test is rejected. Rejection of the stationarity or halfwidth test suggests that additional samples are needed.

Returns a ChainSummary type object with parameters contained in the rows of the value field, and numbers of burn-in sequences to discard, whether the stationarity tests are passed (1 = yes, 0 = no), their p-values ($p > \alpha$ implies stationarity), posterior means, halfwidths of their $(1 - \alpha) 100\%$ estimation intervals, and whether the halfwidth tests are passed (1 = yes, 0 = no) in the columns. Results are chain-specific.

c : sampler output on which to perform calculations.

alpha : significance level for evaluations of stationarity tests and calculations of relative estimation interval halfwidths.

eps : target ratio for the relative halfwidths.

etype : method for computing Monte Carlo standard errors. See mcse() for options.

args... : additional arguments to be passed to the etype method.

x : vector on which to perform calculations.

start : ??

Raftery and Lewis Diagnostic

Main.MCPhylo.rafterydiag — Function

rafterydiag(x::Vector{T}; q::Real=0.025, r::Real=0.005,
                  s::Real=0.95, eps::Real=0.001,
                  range::AbstractRange=1:1:length(x)) where {T<:Real}

rafterydiag(c::AbstractChains; q::Real=0.025, r::Real=0.005,
                 s::Real=0.95, eps::Real=0.001)

Compute the convergence diagnostic of Raftery and Lewis for MCMC sampler output. The diagnostic is designed to determine the number of autocorrelated samples required to estimate a specified quantile $\theta_q$, such that $\Pr(\theta \le \theta_q) = q$, within a desired degree of accuracy. In particular, if $\hat{\theta}_q$ is the estimand and $\Pr(\theta \le \hat{\theta}_q) = \hat{P}_q$ the estimated cumulative probability, then accuracy is specified in terms of r and s, where $\Pr(q - r < \hat{P}_q < q + r) = s$. Thinning may be employed in the calculation of the diagnostic to satisfy its underlying assumptions. However, users may not want to apply the same (or any) thinning when estimating posterior summary statistics because doing so results in a loss of information. Accordingly, sample sizes estimated by the diagnostic tend to be conservative (too large).

Returns a ChainSummary type object with parameters contained in the rows of the value field, and thinning intervals employed, numbers of samples to discard as burn-in sequences, total numbers $(N)$ to burn-in and retain, numbers of independent samples that would be needed $(Nmin)$, and dependence factors $(N / Nmin)$ in the columns. Results are chain-specific.

c : sampler output on which to perform calculations.

q : posterior quantile of interest.

r : margin of error for estimated cumulative probabilities.

s : probability for the margin of error.

eps : tolerance within which the probabilities of transitioning from initial to retained iterations are within the equilibrium probabilities for the chain. This argument determines the number of samples to discard as a burn-in sequence and is typically left at its default value.

x : vector on which to perform calculations.

range : ??

Monte Carlo Standard Errors

Main.MCPhylo.mcse — Function

mcse(x::Vector{T}, method::Symbol=:imse; args...) where {T<:Real}

Compute Monte Carlo standard errors.

Returns the numeric standard error value.

x : time series of values on which to perform calculations.

method : method used for the calculations. Options are

:bm : batch means, with optional argument size::Integer=100 determining the number of sequential values to include in each batch. This method requires that the number of values in x is at least 2 times the batch size.

:imse : initial monotone sequence estimator.

:ipse : initial positive sequence estimator.

args... : additional arguments for the calculation method.

ModelChains

Main.MCPhylo.ModelChains — Type

ModelChains(c::Chains, m::Model)

See Chains().

Model-Based Inference

Distributions.logpdf — Function

logpdf(mc::ModelChains,
        nodekeys::Vector{Symbol}=keys(mc.model, :stochastic))

Distributions.logpdf — Method

logpdf(mc::ModelChains, f::Function, nodekeys::Vector{Symbol})

Compute the sum of log-densities at each iteration of MCMC output for stochastic nodes.

Returns a ModelChains object of resulting summed log-densities at each MCMC iteration of the supplied chain.

mc : sampler output from a model fit with the mcmc()` function.

nodekey/nodekeys : stochastic model node(s) over which to sum densities (default: all).

f : ??

Main.MCPhylo.dic — Method

dic(mc::ModelChains)

Compute the Deviance Information Criterion (DIC) of Spiegelhalter et al. and Gelman et al. from MCMC sampler output.

Returns a ChainSummary type object with DIC results from the methods of Spiegelhalter and Gelman in the first and second rows of the value field, and the DIC value and effective numbers of parameters in the first and second columns; where

$\text{DIC} = -2 \mathcal{L}(\bar{\Theta}) + 2 p,$

such that $\mathcal{L}(\bar{\Theta})$ is the log-likelihood of model outputs given the expected values of model parameters $\Theta$, and $p$ is the effective number of parameters. The latter is defined as $p_D = -2 \bar{\mathcal{L}}(\Theta) + 2 \mathcal{L}(\bar{\Theta})$ for the method of Spiegelhalter and as $p_V = \frac{1}{2} \operatorname{var}(-2 \mathcal{L}(\Theta))$ for the method of Gelman. Results are for all chains combined.

mc : sampler output from a model fit with the mcmc() function.

StatsBase.predict — Function

predict(mc::ModelChains,
         nodekeys::Vector{Symbol}=keys(mc.model, :output))

Generate MCMC draws from a posterior predictive distribution.

Returns a ModelChains object of draws simulated at each MCMC iteration of the supplied chain. For observed data node $y$, simulation is from the posterior predictive distribution

$p(\tilde{y} | y) = \int p(\tilde{y} | \Theta) p(\Theta | y) d\Theta,$

where $\tilde{y}$ is an unknown observation on the node, $p(\tilde{y} | \Theta)$ is the data likelihood, and $p(\Theta | y)$ is the posterior distribution of unobserved parameters $\Theta$.

mc : sampler output from a model fit with the mcmc()` function.

nodekey/nodekeys : observed Stochastic model node(s) for which to generate draws from the predictive distribution (default: all observed data nodes).

Plot

Main.MCPhylo.autocorplot — Method

autocorplot(c::AbstractChains, indeces::Vector{Int64}; maxlag::Integer=round(Int, 10 * log10(length(c.range))), legend::Bool=false, na...)

–- INTERNAL –- Helper function called by the plot function when a autocorrelation plot is needed.

Main.MCPhylo.barplot — Method

barplot(c::AbstractChains, indeces::Vector{Int64};
        legend::Bool=false, position::Symbol=:stack, na...)

–- INTERNAL –- Helper function called by the plot function when a barplot is needed.

Main.MCPhylo.check_vars — Method

checkvars(simnames::Vector{AbstractString}, vars::Vector{String})::Vector{Int64}

–- INTERNAL –- Helper function that returns a list of indeces that correspond to specific variables. Only those variables get plotted in later steps.

Main.MCPhylo.contourplot — Method

contourplot(c::AbstractChains; <keyword arguments>)::Vector{Plots.Plot}

Function that takes a MCMC chain and creates contourplots. If variables are limited with 'vars' keyword argument, at least 2 variables have to be specified, or no contourplot can be drawn.

Arguments

'vars::Vector{String}': specifies the variables of the chain that are plotted

. 'filename::String'="": when given, the plots will be saved to a file

'fmt::Symbol': specifies the format of the output file

'nrow::Integer' / 'ncol::Integer': Define layout of the plot window(s), i.e. how many plots on each page

'legend::Bool=false': Turn plot legend on / off

Main.MCPhylo.contourplot_int — Method

contourplot_int(c::AbstractChains, indeces::Vector{Int64}; bins::Integer=100, na...)

–- INTERNAL –- Helper function called by the contourplot function.

Main.MCPhylo.densityplot — Method

densityplot(c::AbstractChains, indeces::Vector{Int64}; legend::Bool=false, trim::Tuple{Real, Real}=(0.025, 0.975), na...)

–- INTERNAL –- Helper function called by the plot function when a densityplot is needed.

Main.MCPhylo.draw — Method

draw(p::Array{T}; fmt::Symbol=:svg, filename::String="", nrow::Integer=3,
     ncol::Integer=2, byrow::Bool=false, ask::Bool=true) where T<:Plots.Plot

–- INTERNAL –- Helper function for the plot functions. Displays plots and - if wanted - saves them to a file.

Main.MCPhylo.meanplot — Method

meanplot(c::AbstractChains, indeces::Vector{Int64};
         legend::Bool=false, na...)

–- INTERNAL –- Helper function called by the plot function when a meanplot is needed.

Main.MCPhylo.mixeddensityplot — Method

mixeddensityplot(c::AbstractChains, indeces::Vector{Int64};
                 barbounds::Tuple{Real, Real}=(0, Inf), args...)

–- INTERNAL –- Helper function called by the plot function. Checks for each variable if it is discrete or not and plots a barplot for discrete variables and a densityplot for indiscrete variables.

Main.MCPhylo.traceplot — Method

traceplot(c::AbstractChains, indeces::Vector{Int64};
          legend::Bool=false, na...)

–- INTERNAL –- Helper function called by the plot function when a traceplot is needed.

RecipesBase.plot — Function

plot(c::AbstractChains, ptype::Vector{Symbol}=[:trace, :density]; <keyword arguments>)::Array{Plots.Plot}

Function that takes a MCMC chain and creates various different plots (trace & density by default).

Arguments

'vars::Vector{String}': specifies the variables of the chain that are plotted

. 'filename::String'="": when given, the plots will be saved to a file

'fmt::Symbol': specifies the format of the output file

'nrow::Integer' / 'ncol::Integer': Define layout of the plot window(s), i.e. how many plots on each page

'legend::Bool=false': Turn plot legend on / off

'args': Plottype specific arguments, like the number of bins for the contour plot or if the barplots bars should be stacked or not. Check the specific plot functions below to use these arguments.

RecipesBase.plot — Method

plot(c::AbstractChains, ptype::Symbol, indeces::Vector{Int64}; legend::Bool=false, args...)

–- INTERNAL –- Helper function to the actual plot function. Calls the specific plotting functions that are needed.

Posterior Summary Statistics

DataAPI.describe — Method

describe(io::IO, c::AbstractChains;
              q::Vector=[0.025, 0.25, 0.5, 0.75, 0.975], etype=:bm, args...)

Compute summary statistics for MCMC sampler output.

Returns results from calls to summarystats(c, etype, args...) and quantile(c, q) are printed for all chains combined, and a value of nothing is returned.

c : sampler output on which to perform calculations.

q : probabilities at which to calculate quantiles.

etype : method for computing Monte Carlo standard errors. See mcse() for options.

args... : additional arguments to be passed to the etype method.

Main.MCPhylo.changerate — Method

changerate(c::AbstractChains)

Estimate the probability, or rate per iteration, $\Pr(\theta^i \ne \theta^{i-1})$ of a state space change for iterations $i = 2, \ldots, N$ in MCMC sampler output. Estimation is performed for each parameter univariately as well as for the full parameter vector multivariately. For continuous output generated from samplers, like Metropolis-Hastings, whose algorithms conditionally accept candidate draws, the probability can be viewed as the acceptance rate.

Returns a ChainSummary type object with parameters in the rows of the value field, and the estimated rates in the column. Results are for all chains combined.

c : sampler output on which to perform calculations.

Main.MCPhylo.hpd — Method

hpd(c::AbstractChains; alpha::Real=0.05)

Main.MCPhylo.hpd — Method

hpd(x::Vector{T}; alpha::Real=0.05) where {T<:Real}

Compute highest posterior density (HPD) intervals of Chen and Shao [16] for MCMC sampler output. HPD intervals have the desirable property of being the smallest intervals that contain a given probability. However, their calculation assumes unimodal marginal posterior distributions, and they are not invariant to transformations of parameters like central (quantile-based) posterior intervals.

Returns a ChainSummary type object with parameters contained in the rows of the value field, and lower and upper intervals in the first and second columns. Results are for all chains combined.

c : sampler output on which to perform calculations.

alpha : the 100 * (1 - alpha)% interval to compute.

x : vector on which to perform calculations.

Statistics.cor — Method

cor(c::AbstractChains)

Compute cross-correlations for MCMC sampler output.

Returns a ChainSummary type object with the first and second dimensions of the value field indexing the model parameters between which correlations. Results are for all chains combined.

c : sampler output on which to perform calculations.

Statistics.quantile — Method

quantile(c::AbstractChains; q::Vector=[0.025, 0.25, 0.5, 0.75, 0.975])

Compute posterior quantiles for MCMC sampler output.

Returns a ChainSummary type object with parameters contained in the rows of the value field, and quantiles in the columns. Results are for all chains combined.

c : sampler output on which to perform calculations.

q : probabilities at which to compute quantiles.

StatsBase.autocor — Method

autocor(c::AbstractChains; lags::Vector=[1, 5, 10, 50],
         relative::Bool=true)

Compute lag-k autocorrelations for MCMC sampler output.

Returns a ChainSummary type object with model parameters indexed by the first dimension of value, lag-autocorrelations by the second, and chains by the third.

c : sampler output on which to perform calculations.

lags : lags at which to compute autocorrelations.

relative : whether the lags are relative to the thinning interval of the output (true) or relative to the absolute iteration numbers (false).

StatsBase.summarystats — Method

summarystats(c::AbstractChains; etype=:bm, args...)

Compute posterior summary statistics for MCMC sampler output.

Returns a ChainSummary type object with parameters in the rows of the value field; and the sample mean, standard deviation, standard error, Monte Carlo standard error, and effective sample size in the columns. Results are for all chains combined.

c : sampler output on which to perform calculations.

etype : method for computing Monte Carlo standard errors. See mcse() for options.

args... : additional arguments to be passed to the etype method.