Parser

Main.MCPhylo.ParseCSV — Function

ParseCSV(filename::String)

This function parses a CSV file which stores the input for the MCMC compuation. The file should follow the conventions used for MrBayes.

Main.MCPhylo.create_csvdf — Function

function create_csvdf(filecontnt::Array{String}, separator::AbstractString=",")::DataFrame

This function parses a CSV file and returns its content as a DataFrame.

NOTE: May later on be replaced by the respective DataFrames function.

Main.MCPhylo.ParseNewick — Method

ParseNewick(filename::String)

This is the main function, which parses a file, containing Newick strings.

Main.MCPhylo.ParseNexus — Method

ParseNexus(filename::String)

This function parses a NEXUS file which stores the input for the MCMC compuation. The file should follow the conventions used for MrBayes.

Main.MCPhylo.create_nexusdf — Method

create_nexusdf(filecontent::Array{String})::DataFrame

This function creates a DataFrame of the acutal data.

Main.MCPhylo.extract_meta_info — Method

extract_meta_info(content::Array{String})

This function extracts some meta information from the content of the nexus file. It "eats-up" the stack