samplers

Functions found in all files of samplers folder, ordered according to file of origin.

abc.jl

Main.MCPhylo.ABC — Function

          ABC(params::ElementOrVector{Symbol},

          scale::ElementOrVector{T}, summary::Function,

          epsilon::Real; kernel::KernelDensityType=SymUniform,

          dist::Function=(Tsim, Tobs) -> sqrt(sum(abs2, Tsim - Tobs)),

          proposal::SymDistributionType=Normal, maxdraw::Integer=1,

          nsim::Integer=1, decay::Real=1.0, randeps::Bool=false,

          args...) where {T<:Real}

Construct a Sampler object for ABC sampling. Parameters are assumed to be continuous, but may be constrained or unconstrained. Returns a Sampler{ABCTune} type object.

Arguments

params: stochastic node(s) to be updated with the sampler. Constrained parameters are mapped to unconstrained space according to transformations defined by the Stochastic unlist() function.

scale : scaling value or vector of the same length as the combined elements of nodes params for the proposal distribution. Values are relative to the unconstrained parameter space, where candidate draws are generated.

summary : function that takes a vector of observed or simulated data and returns a summary statistic or vector of statistics.

epsilon : target tolerance for determining how similar observed and simulated data summary statistics need to be in order to accept a candidate draw.

kernel : weighting kernel density of type Biweight, Cosine, Epanechnikov, Normal, SymTriangularDist, SymUniform, or Triweight to use in measuring similarity between observed and simulated data summary statistics. Specified epsilon determines the standard deviation of Normal kernels and widths of the others.

dist : positive function for the kernel density to compute distance between vectors of observed (Tobs) and simulated (Tsim) data summary statistics (default: Euclidean distance).

proposal : symmetric distribution of type Biweight, Cosine, Epanechnikov, Normal, SymTriangularDist, SymUniform, or Triweight to be centered around current parameter values and used to generate proposal draws. Specified scale determines the standard deviations of Normal proposals and widths of the others.

maxdraw : maximum number of unaccepted candidates to draw in each call of the sampler. Draws are generated until one is accepted or the maximum is reached. Larger values increase acceptance rates at the expense of longer runtimes.

nsim : number of data sets to simulate in deciding whether to accept a candidate draw. Larger values lead to closer approximations of the target distribution at the expense of longer runtimes.

decay : if 0 < decay <= 1, the rate at which internal tolerances are monotonically decreased from the initial distance between observed and simulated summary statistics toward the maximum of each subsequent distance and epsilon; if decay = 0, internal tolerances are fixed at epsilon.

randeps : whether to perturb internal tolerances by random exponential variates.

args... : additional keyword arguments to be passed to the dist function.

amm.jl

Main.MCPhylo.AMM — Function

AMM(params::ElementOrVector{Symbol}, Sigma::Matrix{T};
    adapt::Symbol=:all, args...) where {T<:Real}

Construct a Sampler object for AMM sampling. Parameters are assumed to be continuous, but may be constrained or unconstrained. Returns a Sampler{AMMTune} type object.

Arguments

params : stochastic node(s) to be updated with the sampler. Constrained parameters are mapped to unconstrained space according to transformations defined by the Stochastic unlist() function.

Sigma : covariance matrix for the non-adaptive multivariate normal proposal distribution. The covariance matrix is relative to the unconstrained parameter space, where candidate draws are generated.

adapt : type of adaptation phase. Options are

:all : adapt proposal during all iterations.

:burnin : adapt proposal during burn-in iterations.

:none : no adaptation (multivariate Metropolis sampling with fixed proposal).

args... : additional keyword arguments to be passed to the AMMVariate constructor.

StatsBase.sample! — Method

sample!(v::AMMVariate, logf::Function; adapt::Bool=true)

Draw one sample from a target distribution using the AMM sampler. Parameters are assumed to be continuous and unconstrained. Returns $v$ updated with simulated values and associated tuning parameters.

Arguments

v : current state of parameters to be simulated. When running the sampler in adaptive mode, the v argument in a successive call to the function will contain the tune field returned by the previous call.

adapt : whether to adaptively update the proposal distribution.

amwg.jl

Main.MCPhylo.AMWG — Function

AMWG(params::ElementOrVector{Symbol},

sigma::ElementOrVector{T};

adapt::Symbol=:all,

args...) where {T<:Real}

Construct a Sampler object for AMWG sampling. Parameters are assumed to be continuous, but may be constrained or unconstrained. Returns a Sampler{ABCTune} type object.

Arguments

params: stochastic node(s) to be updated with the sampler. Constrained parameters are mapped to unconstrained space according to transformations defined by the Stochastic unlist() function.

sigma: scaling value or vector of the same length as the combined elements of nodes

params, defining initial standard deviations for univariate normal proposal distributions. Standard deviations are relative to the unconstrained parameter space, where candidate draws are generated.

adapt : type of adaptation phase. Options are

:all : adapt proposal during all iterations.

:burnin : adapt proposal during burn-in iterations.

:none : no adaptation (Metropolis-within-Gibbs sampling with fixed proposal).

args...: additional keyword arguments to be passed to the AMWGVariate constructor.

StatsBase.sample! — Method

sample!(v::AMWGVariate, logf::Function; adapt::Bool=true)

Draw one sample from a target distribution using the AMWG sampler. Parameters are assumed to be continuous and unconstrained. Returns v updated with simulated values and associated tuning parameters.

Arguments

v: current state of parameters to be simulated. When running the sampler in adaptive mode, the v argument in a successive call to the function will contain the tune field returned by the previous call.

adapt: whether to adaptively update the proposal distribution.

bhmc.jl

Main.MCPhylo.BHMC — Function

BHMC(params::ElementOrVector{Symbol}, traveltime::Real)

Construct a Sampler object for BHMC sampling. Parameters are assumed to have binary numerical values (0 or 1).

Returns a Sampler{BHMCTune} type object.

Arguments

params: stochastic node(s) to be updated with the sampler.

traveltime: length of time over which particle paths are simulated. It is recommended that supplied values be of the form $(n + \frac{1}{2}) \pi$, where optimal choices of n $\in \mathbb{Z}^+$ are expected to grow with the parameter space dimensionality.

StatsBase.sample! — Method

sample!(v::BHMCVariate, logf::Function)

Draw one sample from a target distribution using the BHMC sampler. Parameters are assumed to have binary numerical values (0 or 1). Returns v updated with simulated values and associated tuning parameters.

Arguments

v: current state of parameters to be simulated.

bia.jl

Main.MCPhylo.BIA — Function

BIA(params::ElementOrVector{Symbol}; args...)

Construct a Sampler object for BIA sampling. Parameters are assumed to have binary numerical values (0 or 1).

Returns a Sampler{BIATune} type object.

Arguments

params: stochastic node(s) to be updated with the sampler.

args...: additional keyword arguments to be passed to the BIAVariate constructor.

StatsBase.sample! — Method

sample!(v::BIAVariate, logf::Function)

Draw one sample from a target distribution using the BIA sampler. Parameters are assumed to have binary numerical values (0 or 1).

Returns v updated with simulated values and associated tuning parameters.

Arguments

v: current state of parameters to be simulated.

bmc3.jl

Main.MCPhylo.BMC3 — Function

BMC3(params::ElementOrVector{Symbol}; k::F=1) where {F<:BMC3Form}

Construct a Sampler object for BMC3 sampling. Parameters are assumed to have binary numerical values (0 or 1).

Returns a Sampler{BMC3Tune{typeof(k)}} type object.

Arguments

params: stochastic node(s) to be updated with the sampler.

k: number of parameters or vector of parameter indices to select at random for simultaneous updating in each call of the sampler.

StatsBase.sample! — Method

sample!(v::SamplerVariate{BMC3Tune{F}}, logf::Function) where {F<:BMC3Form}

Draw one sample from a target distribution using the BMC3 sampler. Parameters are assumed to have binary numerical values (0 or 1).

Returns v updated with similar values and associated tuning parameters.

Arguments

v: current state of parameters to be simulated.

bmg.jl

Main.MCPhylo.BMG — Function

BMG(params::ElementOrVector{Symbol}; k::F=1) where {F<:BMGForm}

Construct a Sampler object for BMG sampling. Parameters are assumed to have binary numerical values (0 or 1).

Returns a Sampler{BMGTune{typeof(k)}} type object.

Arguments

params: stochastic node(s) to be updated with the sampler.

k: number of parameters or vector of parameter indices to select at random for simultaneous updating in each call of the sampler.

StatsBase.sample! — Method

sample!(v::SamplerVariate{BMGTune{F}}, logf::Function) where {F<:BMGForm}

Draw one sample from a target distribution using the BMG sampler. Parameters are assumed to have binary numerical values (0 or 1).

Returns v updated with simulated values and associated tuning parameters.

Arguments

v: current state of parameters to be simulated.

dgs.jl

Main.MCPhylo.DGS — Function

DGS(params::ElementOrVector{Symbol})

Construct a Sampler object for which DGS sampling is to be applied separately to each of the supplied parameters. Parameters are assumed to have discrete univariate distributions with finite supports.

Returns a Sampler{DSTune{Function}} type object.

Arguments

params: stochastic node(s) to be updated with the sampler.

StatsBase.sample! — Method

sample!(v::DGSVariate, mass::Function)

Draw one sample directly from a target probability mass function. Parameters are assumed to have discrete and finite support.

Returns v updated with simulated values and associated tuning parameters.

StatsBase.sample! — Method

sample!(v::DiscreteVariate, mass::Vector{Float64})

Draw one sample directly from a target probability mass function. Parameters are assumed to have discrete and finite support.

Returns v updated with simulated values and associated tuning parameters.

Arguments

v: current state of parameters to be simulated.

hmc.jl

Main.MCPhylo.HMC — Function

HMC(params::ElementOrVector{Symbol}, epsilon::Real, L::Integer; args...)

Construct a Sampler object for HMC sampling. Parameters are assumed to be continuous, but may be constrained or unconstrained.

Returns a Sampler{HMCTune} type object.

HMC(params::ElementOrVector{Symbol}, epsilon::Real, L::Integer, Sigma::Matrix{T}; args...)

Construct a Sampler object for HMC sampling. Parameters are assumed to be continuous, but may be constrained or unconstrained.

Returns a Sampler{HMCTune} type object.

Arguments

params: stochastic node(s) to be updated with the sampler. Constrained parameters are mapped to unconstrained space according to transformations defined by the Stochastic unlist() function.

epsilon: step size.

L: number of steps to take in the Leapfrog algorithm.

Sigma: covariance matrix for the multivariate normal proposal distribution. The covariance matrix is relative to the unconstrained parameter space, where candidate draws are generated. If omitted, the identity matrix is assumed.

dtype : differentiation for gradient calculations. Options are

:central : central differencing

:forward : forward differencing.

StatsBase.sample! — Method

sample!(v::HMCVariate, logfgrad::Function)

Draw one sample from a target distribution using the HMC sampler. Parameters are assumed to be continuous and unconstrained.

Returns v updated with simulated values and associated tuning parameters.

Arguments

v: current state of parameters to be simulated.

mala.jl

Main.MCPhylo.MALA — Function

MALA(params::ElementOrVector{Symbol}, epsilon::Real; args...)

Construct a Sampler object for MALA sampling. Parameters are assumed to be continuous, but may be constrained or unconstrained.

Returns a Sampler{MALATune}` type object.

MALA(params::ElementOrVector{Symbol}, epsilon::Real, Sigma::Matrix{T}; args...)

Construct a Sampler object for MALA sampling. Parameters are assumed to be continuous, but may be constrained or unconstrained.

Returns a Sampler{MALATune}` type object.

Arguments

params: stochastic node(s) to be updated with the sampler. Constrained parameters are mapped to unconstrained space according to transformations defined by the Stochastic unlist() function.

epsilon: factor by which the drift and covariance matrix of the proposal distribution are scaled.

Sigma: covariance matrix for the multivariate normal proposal distribution. The covariance matrix is relative to the unconstrained parameter space, where candidate draws are generated. If omitted, the identity matrix is assumed.

dtype : differentiation for gradient calculations. Options are

:central : central differencing

:forward : forward differencing.

StatsBase.sample! — Method

sample!(v::MALAVariate, logfgrad::Function)

Draw one sample from a target distribution using the MALA sampler. Parameters are assumed to be continuous and unconstrained.

Returns v updated with simulated values and associated tuning parameters.

Arguments

v: current state of parameters to be simulated.

miss.jl

Main.MCPhylo.MISS — Function

MISS(params::ElementOrVector{Symbol})

Construct a Sampler object to sampling missing output values. The constructor should only be used to sample stochastic nodes upon which no other stochastic node depends. So-called ‘output nodes’ can be identified with the keys() function. Moreover, when the MISS constructor is included in a vector of Sampler objects to define a sampling scheme, it should be positioned at the beginning of the vector. This ensures that missing output values are updated before any other samplers are executed.

Returns a Sampler{Dict{Symbol, MISSTune}} type object.

Arguments

params: stochastic node(s) that contain missing values (NaN) to be updated with the sampler.

nuts.jl

Main.MCPhylo.NUTS — Function

NUTS(params::ElementOrVector{Symbol}; dtype::Symbol=:forward, args...)

Construct a Sampler object for NUTS sampling, with the algorithm’s step size parameter adaptively tuned during burn-in iterations. Parameters are assumed to be continuous, but may be constrained or unconstrained.

Returns a Sampler{NUTSTune} type object.

Arguments

params: stochastic node(s) to be updated with the sampler. Constrained parameters are mapped to unconstrained space according to transformations defined by the Stochastic unlist() function.

dtype : differentiation for gradient calculations. Options are

:central : central differencing

:forward : forward differencing.

args...: additional keyword arguments to be passed to the NUTSVariate constructor.

StatsBase.sample! — Method

sample!(v::NUTSVariate, logfgrad::Function; adapt::Bool=false)

Draw one sample from a target distribution using the NUTS sampler. Parameters are assumed to be continuous and unconstrained.

Returns v updated with simulated values and associated tuning parameters.

Arguments

v: current state of parameters to be simulated. When running the sampler in adaptive mode, the v argument in a successive call to the function will contain the tune field returned by the previous call.

adapt: whether to adaptively update the epsilon step size parameter.

rwm.jl

Main.MCPhylo.RWM — Method

RWM(params::ElementOrVector{Symbol},
              scale::ElementOrVector{T}; args...) where {T<:Real})

Construct a Sampler object for RWM sampling. Parameters are assumed to be continuous, but may be constrained or unconstrained.

Returns a Sampler{RWMTune} type object.

Arguments

params: stochastic node(s) to be updated with the sampler. Constrained parameters are mapped to unconstrained space according to transformations defined by the Stochastic unlist() function.

scale: scaling value or vector of the same length as the combined elements of nodes params for the proposal distribution. Values are relative to the unconstrained parameter space, where candidate draws are generated.

args...: additional keyword arguments to be passed to the RWMVariate constructor.

StatsBase.sample! — Method

sample!(v::RWMVariate, logf::Function)

Draw one sample from a target distribution using the RWM sampler. Parameters are assumed to be continuous and unconstrained.

Returns v updated with simulated values and associated tuning parameters.

Arguments

v: current state of parameters to be simulated.

slice.jl

Main.MCPhylo.Slice — Function

Slice(params::ElementOrVector{Symbol}, width::ElementOrVector{T}, ::Type{F}=Multivariate; transform::Bool=false) where {T<:Real, F<:SliceForm}

Construct a Sampler object for Slice sampling. Parameters are assumed to be continuous, but may be constrained or unconstrained.

Returns a Sampler{SliceTune{Univariate}} or Sampler{SliceTune{Multivariate}} type object if sampling univariately or multivariately, respectively.

Arguments

params: stochastic node(s) to be updated with the sampler.

width: scaling value or vector of the same length as the combined elements of nodes params, defining initial widths of a hyperrectangle from which to simulate values.

F : sampler type. Options are

:Univariate : sequential univariate sampling of parameters.

:Multivariate : joint multivariate sampling.

transform: whether to sample parameters on the link-transformed scale (unconstrained parameter space). If true, then constrained parameters are mapped to unconstrained space according to transformations defined by the Stochastic unlist() function, and width is interpreted as being relative to the unconstrained parameter space. Otherwise, sampling is relative to the untransformed space.

slicesimplex.jl

Main.MCPhylo.SliceSimplex — Function

SliceSimplex(params::ElementOrVector{Symbol}; args...)

Construct a Sampler object for which SliceSimplex sampling is to be applied separately to each of the supplied parameters. Parameters are assumed to be continuous and constrained to a simplex.

Returns a Sampler{SliceSimplexTune} type object.

Arguments

params: stochastic node(s) to be updated with the sampler.

args...: additional keyword arguments to be passed to the SliceSimplexVariate constructor.

StatsBase.sample! — Method

sample!(v::SliceSimplexVariate, logf::Function)

Draw one sample from a target distribution using the SliceSimplex sampler. Parameters are assumed to be continuous and constrained to a simplex.

Returns v updated with simulated values and associated tuning parameters.

Arguments

v: current state of parameters to be simulated.