Tree Functionality

Main.MCPhylo.from_df — Method

from_df(df::DataFrame)::Node

This function takes a DataFrame and turns it into a tree. It assumes a rooted binary tree is stored in the matrix. No checks are performed.

Main.MCPhylo.to_covariance_ultra — Method

to_covariance_ultra(tree::Node)::Array{T,2} where T<: Real

Get the covariance matrix of the ultrametric version of tree with height 1.

Main.MCPhylo.check_binary — Method

check_binary(root::Node)::Bool
checks to see if given tree is binary; returns true if properly formatted and false otherwise

Main.MCPhylo.create_tree_from_leaves_bin — Method

create_tree_from_leaves(leaf_nodes::Vector{T})::Node

This function creates a random binary tree from a list of leaf nodes. The root node as access point for the tree is returned.

Main.MCPhylo.force_ultrametric! — Method

force_ultrametric!(root::T) where T<:GeneralNode

Force an ultrametric version of the tree.

Main.MCPhylo.get_mother — Method

function get_mother(node::T)::T  where T<:GeneralNode

This function gets the mother of node. It does so by looking for the respective binary representation of the mother node.

Main.MCPhylo.get_sum_seperate_length! — Method

get_sum_seperate_length!(root::Node)::Vector{Float64}

This function gets the sum of the branch lengths of the internal branches and the branches leading to the leave nodes.

Main.MCPhylo.get_sum_seperate_length! — Method

get_sum_seperate_length!(root::Node)::Vector{Float64}

This function gets the sum of the branch lengths of the internal branches and the branches leading to the leave nodes.

Main.MCPhylo.node_height — Method

node_height(root::T, mv::Float64)::Float64  where T<:GeneralNode

Calculate the height of a node.

Main.MCPhylo.number_nodes! — Method

number_nodes!(root::T)::Nothing  where T<:GeneralNode

This function assigns a unique, sequential number to each node.

Main.MCPhylo.random_node — Method

function random_node(root::T)::T  where T<:GeneralNode

This function returns a random node from the tree.

Main.MCPhylo.set_binary! — Method

set_binary!(root::Node)

Assign a binary representation to each node, which specifies the path from the root to this node via the binary representation of the node. A left turn is a 1 in binary and a right turn a 0.

Main.MCPhylo.neighbor_joining_int — Method

neighbor_joining_int(dm::Array{Float64,2},leaves::Vector{Node})

Internal function that is called by both neighbor_joining methods. Contains the actual neighbor-joining algorithm, i.e. builds a phylogenetic tree from the given distance matrix and array of leaves. Returns a node of that tree, from which it can be traversed.

Main.MCPhylo.upgma_int — Method

upgma_int(dm::Array{Float64,2},leaves::Vector{Node})

Internal function that is called by both UPGMA methods. Contains the actual UPGMA algorithm, i.e. builds a phylogenetic tree from the given distance matrix and array of leaves. Returns a node of that tree, from which it can be traversed.

Main.MCPhylo.find_common_clusters — Method

find_common_clusters(ref_tree, tree:T)
    ::Dict{Int64, Tuple{Bool, Union{Float64, Missing}}}

Use Day's algorithm to create a dictionary, that tells us for each node of the second input tree, if its corresponding cluster is a common cluster of the trees

Main.MCPhylo.max_leaf_rank — Method

max_leaf_rank(leaf_ranks::Dict{String, Int64}, node::T)
    ::Int64 where T<:AbstractNode

Recursive helper function to find the highest ranked leaf descendant of a node

Main.MCPhylo.merge_trees! — Method

merge_trees!(ref_tree::T, tree::T)::Tuple{T, Vector{T}} where T<:AbstractNode

Merge two compatible trees, i.e. inserts all cluster of the first tree, which aren't already in the second tree, into the secon tree

Main.MCPhylo.min_leaf_rank — Method

min_leaf_rank(leaf_ranks::Dict{String, Int64}, node::T)
    ::Int64 where T <: AbstractNode

Recursive helper function to find the lowest ranked leaf descendant of a node

Main.MCPhylo.one_way_compatible — Method

one_way_compatible(ref_tree::T, tree::T)::T where T<:AbstractNode

Takes two trees and returns a copy of the first one, where all the clusters that are not compatible with the second tree are removed.

Main.MCPhylo.order_tree! — Method

order_tree!(root::T, cluster_start_indeces::Dict{T, Int64}, leaves=Vector{T}())
    ::Vector{T} where T<:AbstractNode

Helper function to order a tree based on cluster indeces and return the leaves of the ordered tree

Main.MCPhylo.x_left — Method

x_left(node::T)::Tuple{T,Vector{T}} where T<:AbstractNode

Helper function to find ancestor of a leaf that has said leaf as leftmost descendant. Also returns the path from the leaf to the mother of that node.

Main.MCPhylo.x_right — Method

x_right(node::T)::Tuple{T,Vector{T}} where T<:AbstractNode

Helper function to find ancestor of a leaf that has said leaf as rightmost descendant. Also returns the path from the leaf to the mother of that node.

Main.MCPhylo.get_bipartitions — Method

get_bipartitions(tree::T)::Vector{Tuple} where T <:GeneralNode

Get a vector of all bipartions of tree. The resulting vector contains Tuples of sets representing the bipartions.

Main.MCPhylo.change_edge_length! — Method

change_edge_length!(root::T) where T <:GeneralNode

Pick a random node and increase or decrease its length randomly.

Main.MCPhylo.move! — Method

move!(node1::Node, node2::Node, proportion::Float64)

Change the incomming length of node1 and node2 while keeping there combined length constant.

Main.MCPhylo.perform_spr — Method

perform_spr(root::Node)

performs SPR on binary tree

Main.MCPhylo.risky_SPR! — Method

    risky_SPR!(root::Node)::AbstractNode
Performs SPR on tree in place; takes reference to root of tree, boolean value necessary to determine if tree should be treated as binary or not
Returns reference to root of altered tree. Does not check for correct formatting of tree.

Main.MCPhylo.risky_SPR — Method

    risky_SPR(root::Node)::AbstractNode
Performs SPR on tree in place; takes reference to root of tree, boolean value necessary to determine if tree should be treated as binary or not
Returns copy of root of altered tree. Does not check for correct formatting of tree.

Main.MCPhylo.find_binary — Method

find_binary(root::T, bin::String)::T where T<:GeneralNode

Find a node by its binary representation. The function assumes that the node is present in the tree.

Do not use this function if you are unsure wheter the node is in the tree at all.

Main.MCPhylo.find_num — Method

find_num(root::T, num::Int64, rn::Vector{T})::Bool  where T<:GeneralNode

Do a post order traversal to find the node corresponding to the num.

Main.MCPhylo.find_num — Method

find_num(root::T, num::Int64)  where T<:GeneralNode

Find a node by its number. The function assumes that the node is present in the tree.

Do not use this function if you are unsure wheter the node is in the tree at all.

Main.MCPhylo.level_traverse — Method

level_traverse(node::T, level::Int64, stack::Array{T})::Bool where T <:GeneralNode

This function traverses a level of the tree specified through node. The level is specified via the level argument and the nodes visited are stored in the stack. This function is intended as the internal worker for the level_order function.