Likelihood

Main.MCPhylo.FelsensteinFunction — Method

FelsensteinFunction(treepostorder::Vector{N}, pi::T, rates::Vector{Float64}, data::Array{Float64,3}, n_c::Int64, blv::Vector{Float64}) where {T<:Real, N<:GeneralNode}

This function calculates the log-likelihood of an evolutiuonary model using the Felsensteins pruning algorithm.

The function is written such that it is differentiable by Zygote 0.5.3.

Main.MCPhylo.BirthDeath — Type

Strict Molecular Clock - BirthDeath Implemented following Yang & Rannala 1997 doi.org/10.1093/oxfordjournals.molbev.a025811

Main.MCPhylo.BirthDeathSimplified — Type

Strict Molecular Clock - Simplified Birth Death Implemented folloing Yang & Rannala 1996 doi.org/10.1007/BF02338839