Model

The Model type is designed to store the set of all model nodes, including parameter set \Theta. In particular, it stores Dependent type objects in its nodes dictionary field. Valid models are ones whose nodes form directed acyclic graphs (DAGs). Sampling functions $\{f_j\}_{j=1}^{B}$ are saved as Sampler objects in the vector of field samplers. Vector elements j=1,\ldots,B correspond to sampling blocks $\{\Theta_j\}_{j=1}^{B}$.

Dependent

MCPhylo.Logical — Type

Logical(f::Function, monitor::Union{Bool, Vector{Int}}=true)

Constructor for a Logical model node. This function assumes a the output of the logical operation to be scalar.

f: Function specifying the deterministic operation performed on its arguments. These arguments are other nodes of the model.

monitor : Indicates whether the results should be monitored, i.e.saved.

Logical(d::Integer, f::Function, monitor::Union{Bool, Vector{Int}}=true)

Constructor for a Logical model node.

d : Specifies the dimension of the output.

f : Specifies the deterministic operation performed on its arguments. These arguments are other nodes of the model.

monitor : Indicates whether the results should be monitored, i.e.saved.

Logical(d::T, f::Function, monitor::Union{Bool, Vector{Int}}=true) where T<:GeneralNode

Constructor for a Logical model node, which can hold a Node structure, i.e. a tree.

f is a function specifying the deterministic operation performed on its arguments. These arguments are other nodes of the model.

monitor indicates whether the results should be monitored, i.e. saved.

MCPhylo.Stochastic — Type

Stochastic(f::Function, monitor::Union{Bool, Vector{Int}}=true)

Constructor for a Stochastic model node. This function assumes a the output of the logical operation to be scalar.

f : Specifies the distributional relationship between the arguments and the node. These arguments are other nodes of the model.

monitor : Indicates whether the results should be monitored, i.e. saved.

Stochastic(d::N, f::Function, monitor::Union{Bool, Vector{Int}}=true) where N<:GeneralNode

Constructor for a Stochastic model node, which can hold a Node structure, i.e. a tree.

f : Specifies the distributional reslationship between the arguments and the node.

These arguments are other nodes of the model.

monitor : Indicates whether the results should be monitored, i.e. saved.

MCPhylo.setinits! — Function

setinits!(l::AbstractLogical, m::Model, ::Any=nothing)

Set initial values for a logical node.

Returns the result of a call to setmonitor!(l, l.monitor) or setmonitor!(d, d.monitor).

l : logical node to which to assign initial values.

m : model containing the node.

setinits!(d::TreeLogical, m::Model, x::T) where {T<:GeneralNode}

Set initial values for a logical node.

Returns the result of a call to setmonitor!(l, l.monitor) or setmonitor!(d, d.monitor).

l : logical node to which to assign initial values.

m : model containing the node.

setinits!(s::ScalarStochastic, m::Model, x::Real)

Set initial values for a stochastic node.

Returns the node with its assigned initial values.

s : stochastic node to which to assign initial values.

m : model containing the node.

x : values to assign to the node.

setinits!(s::ArrayStochastic, m::Model, x::DenseArray)

Set initial values for a stochastic node.

Returns the node with its assigned initial values.

s : stochastic node to which to assign initial values.

m : model containing the node.

x : values to assign to the node.

setinits!(d::TreeStochastic, m::Model, x::T) where {T<:GeneralNode}

Set initial values for a stochastic node.

Returns the node with its assigned initial values.

s or d : stochastic node to which to assign initial values.

m : model containing the node.

x : values to assign to the node.

setinits!(m::Model, inits::Dict{Symbol})

Set the initial values of stochastic model nodes.

Returns the model with stochastic nodes initialized and the iter field set equal to 0.

m : model with nodes to be initialized.

inits : initial values for stochastic model nodes. Dictionary keys and values should be given for each stochastic node.

setinits!(m::Model, inits::Vector{V} where V<:Dict{Symbol})

MCPhylo.setmonitor! — Function

setmonitor!(d::AbstractDependent, monitor::Bool)

Specify node elements to be included in monitored MCMC sampler output.

Returns d with its monitor field updated to reflect the specified monitoring.

d : node whose elements contain sampled MCMC values.

monitor : boolean indicating whether all elements are monitored.

setmonitor!(d::AbstractDependent, monitor::Vector{Int})

Specify node elements to be included in monitored MCMC sampler output.

Returns d with its monitor field updated to reflect the specified monitoring.

d : node whose elements contain sampled MCMC values.

monitor : vector of element-wise indices of elements to monitor.

setmonitor!(d::TreeVariate, monitor::Bool)

Specify node elements to be included in monitored MCMC sampler output.

Returns d with its monitor field updated to reflect the specified monitoring.

d : node whose elements contain sampled MCMC values.

monitor : boolean indicating whether all elements are monitored.

setmonitor!(d::TreeVariate, monitor::Vector{Int})

Specify node elements to be included in monitored MCMC sampler output.

Returns d with its monitor field updated to reflect the specified monitoring.

d : node whose elements contain sampled MCMC values.

monitor : vector of element-wise indices of elements to monitor.

MCPhylo.update! — Function

update!(l::AbstractLogical, m::Model)

Update the values of a logical node according to its relationship with others in a model.

Returns the node with its values updated.

l : logical node to update.

m : model containing the node.

update!(s::AbstractStochastic, m::Model)

Update the values of a stochastic node according to its relationship with others in a model.

Returns the node with its values updated.

s : stochastic node to update.

m : model containing the node.

update!(m::Model, block::Integer=0)

update!(m::Model, nodekeys::Vector{Symbol})

Update values of logical and stochastic model node according to their relationship with others in a model.

Returns the model with updated nodes.

m : mode with nodes to be updated.

block : sampling block of nodes to be updated (default: all blocks).

nodekeys : nodes to be updated in the given order.

Graph

MCPhylo.draw — Function

draw(m::Model; filename::AbstractString="")

Draw a GraphViz DOT-formatted graph representation of model nodes and their relationships.

The model drawn to an external file or standard output. Stochastic, logical, and input nodes will be represented by ellipses, diamonds, and rectangles, respectively. Nodes that are unmonitored in MCMC simulations will be gray-colored.

m : model for which to construct a graph.

filename : external file to which to save the resulting graph, or an empty string to draw to standard output (default). If a supplied external file name does not include a dot (.), the file extension .dot will be appended automatically.

MCPhylo.graph2dot — Function

graph2dot(m::Model)

Draw a GraphViz DOT-formatted graph representation of model nodes and their relationships.

A character string representation of the graph suitable for in-line processing. Stochastic, logical, and input nodes will be represented by ellipses, diamonds, and rectangles, respectively. Nodes that are unmonitored in MCMC simulations will be gray-colored.

m : model for which to construct a graph.

Initialization

MCPhylo.SimulationParameters — Method

SimulationParameters(; burnin::Int64=0, thin::Int64=1, chains::Int64=1,
                     verbose::Bool=true, trees::Bool=false,
                     asdsf::Bool=false, freq::Int64=50,
                     min_splits::Float64=0.1)::SimulationParameters

Construct a SimulationParameters object that defines the simulation parameters for a MCMC simulation.

Returns a SimulationParameters type object.

burnin: controls how many trees are discarded before saving

thin: controls thinning of saved trees

chains: controls how many chains there are

verbose: controls if sampler output is printed to the console

trees: controls if trees should be created during simulation

asdsf: controls if ASDSF should be calculated

freq: controls at which interval trees are used for ASDSF

min_splits: controls the default minimal splits threshold

MCPhylo.setinits! — Method

setinits!(m::Model, inits::Dict{Symbol})

Set the initial values of stochastic model nodes.

Returns the model with stochastic nodes initialized and the iter field set equal to 0.

m : model with nodes to be initialized.

inits : initial values for stochastic model nodes. Dictionary keys and values should be given for each stochastic node.

MCPhylo.setinits! — Method

setinits!(m::Model, inits::Vector{V} where V<:Dict{Symbol})

MCPhylo.setinputs! — Method

setinputs!(m::Model, inputs::Dict{Symbol})

Set the values of input model nodes.

m : model with input nodes to be assigned.

inputs : values for input model nodes. Dictionary keys and values should be given for each input node.

Returns the model with values assigned to input nodes.

MCPhylo.setsamplers! — Method

function setsamplers!(m::Model, samplers::Vector{V} where V<:Sampler)

Set the samplers for the Stocastic nodes of a given Model.

Returns the Model with updated samplers.

m : Model to update.

samplers : block-specific sampelrs.

MCMC

MCPhylo.mcmc — Function

mcmc(mc::ModelChains, iters::Integer; verbose::Bool=true, trees::Bool=false)

This function simulates additional draws from a model.

mc is the results of a previous call to the mcmc function.

iters indicates the number of draws to simulate.

verbose controls whether to print progress statements to the console.

trees indicates if the states of the model nodes describing tree structures should be stored as well.

mcmc(m::Model, inputs::Dict{Symbol},
          inits::Vector{V} where V<:Dict{Symbol},
          iters::Integer; burnin::Integer=0, thin::Integer=1,
          chains::Integer=1, verbose::Bool=true, trees::Bool=false,
          params::SimulationParameters=SimulationParameters())

Simulate MCMC draws from the model m.

inputs is a dictionary storing the values for input model nodes. Dictionary keys and values should be given for each input node.

inits contains dictionaries with initial values for stochastic model nodes. Dictionary keys and values should be given for each stochastic node. Consecutive runs of the simulator will iterate through the vector’s dictionary elements.

iter Specifies the number of draws to generate for each simulation run

burnin specifies the number of initial draws to discard as a burn-in sequence to allow for convergence.

thin is the step-size between draws to store in the output.

chains specifies the number of simulation runs to perform.

verbose indicates whether the sampler progress should be printed at the console.

trees indicates if the states of the model nodes describing tree structures should be stored as well.

params pass one Struct to set all simulation parameters, instead of setting each individually

Model

MCPhylo.Model — Type

Model(; iter::Integer=0, burnin::Integer=0,
  samplers::Vector{Sampler}=Sampler[], nodes...)

Construct a Model object that defines a model for MCMC simulation.

Returns a Model type object.

iter: current iteration of the MCMC simulation.

burnin: number of initial draws to be discarded as a burn-in sequence to allow for convergence.

samplers: block-specific sampling functions.

nodes...: arbitrary number of user-specified arguments defining logical and stochastic nodes in the model. Argument values must be Logical or Stochastic type objects. Their names in the model will be taken from the argument names.

Base.keys — Function

Base.keys(m::Model, ntype::Symbol, at...)

Extract the symbols (keys) for all existing nodes or for nodes of a specified type.

m : model containing the nodes of interest.

ntype : type of nodes to return. Options are

:all : all input, logical, and stochastic model nodes.

:assigned : nodes that have been assigned values.

:block : stochastic nodes being updated by the sampling block(s) at::Integer=0 (default: all blocks).

:dependent : logical and stochastic (dependent) nodes in topologically sorted order.

:independent or :input : input (independent) nodes.

:logical : logical nodes.

:monitor : stochastic nodes being monitored in MCMC sampler output.

:output : stochastic nodes upon which no other stochastic nodes depend.

:source : nodes upon which the node at::Symbol or vector of nodes at::Vector{Symbol} depends.

:stochastic : stochastic nodes.

:target : topologically sorted nodes that depend on the sampling block(s) at::Integer=0 (default: all blocks), node at::Symbol , or vector of nodes at::Vector{Symbol} .

at... : additional positional arguments to be passed to the ntype options, as described above.

Base.keys(c::AbstractChains)

Returns names of parameter elements.

c : Chain to return names of.

Base.show — Function

Base.show(io::IO, m::Model)

Write a text representation of the model, nodes, and attributes to the current output stream.

Base.show(io::IO, c::AbstractChains)

Prints header and values of Chain.

io : IO stream on which to print.

AbstractChains : Chain to print.

MCPhylo.showall — Function

showall(io::IO, m::Model)

Write a verbose text representation of the model, nodes, and attributes to the current output stream.

Simulation

MCPhylo.gettune — Function

gettune(m::Model, block::Integer)

gettune(m::Model)

Get block-sampler tuning parameters.

Returns a Vector{Any} of all block-specific tuning parameters without block input, and turning parameters for the specified block otherwise.

MCPhylo.settune! — Function

settune!(m::Model, tune, block::Integer)

settune!(m::Model, tune::Vector{Any})

Set tuning parameters for one or all blocks.

Assigns desired tune values to model.

m : model containing the nodes of interest.

tune : tune values to be assigned to models; if no block value is input, tune must be a Vector with length equal to m.samplers.

block : Integer denoting which block's tune value is to be reassigned.

Distributions.gradlogpdf — Function

gradlogpdf(m::Model, block::Integer=0, transform::Bool=false;
                dtype::Symbol=:forward)

gradlogpdf(m::Model, targets::Array{Symbol, 1})::Tuple{Float64, Array{Float64}}

Compute the gradient of log-densities for stochastic nodes.

Returns the resulting gradient vector. Method gradlogpdf!() additionally updates model m with supplied values x.

m : model containing the stochastic nodes for which to compute the gradient.

block : sampling block of stochastic nodes for which to compute the gradient (default: all stochastic nodes).

x: value (possibly different than the current one) at which to compute the gradient.

transform: whether to compute the gradient of block parameters on the link–transformed scale.

dtype : type of differentiation for gradient calculations. Options are

:central : central differencing.

:forward : forward differencing.

MCPhylo.gradlogpdf! — Function

gradlogpdf!(m::Model, x::AbstractVector{T}, block::Integer=0,
                  transform::Bool=false; dtype::Symbol=:forward) where {T<:Real}

gradlogpdf!(m::Model, x::AbstractArray{T}, block::Integer=0,transform::Bool=false)
 where T<:GeneralNode

gradlogpdf!(m::Model, x::N, block::Integer=0,transform::Bool=false)::Tuple{Float64, Vector{Float64}}
    where N<:GeneralNode

Returns the resulting gradient vector. Method gradlogpdf!() additionally updates model m with supplied values x.

Distributions.logpdf — Function

logpdf(m::Model, block::Integer=0, transform::Bool=false)

logpdf(m::Model, nodekeys::Vector{Symbol}, transform::Bool=false)

Compute the sum of log-densities for stochastic nodes.

Returns the resulting numeric value of summed log-densities.

m: model containing the stochastic nodes for which to evaluate log-densities.

block : sampling block of stochastic nodes over which to sum densities (default: all stochastic nodes).

nodekeys : nodes over which to sum densities.

x : value (possibly different than the current one) at which to evaluate densities.

transform : whether to evaluate evaluate log-densities of block parameters on the link–transformed scale.

logpdf(m::Model, x::AbstractArray{T}, block::Integer=0,
        transform::Bool=false) where {T<:Real}

logpdf(mc::ModelChains, f::Function, nodekeys::Vector{Symbol})

Compute the sum of log-densities at each iteration of MCMC output for stochastic nodes.

Returns a ModelChains object of resulting summed log-densities at each MCMC iteration of the supplied chain.

mc : sampler output from a model fit with the mcmc()` function.

nodekey/nodekeys : stochastic model node(s) over which to sum densities (default: all).

f : ??

logpdf(mc::ModelChains,
       nodekeys::Vector{Symbol}=keys(mc.model, :stochastic))

Distributions.logpdf! — Function

logpdf!(m::Model, x::N, block::Integer=0,
          transform::Bool=false) where N<:GeneralNode

StatsBase.sample! — Method

sample!(m::Model, block::Integer=0)

Generate one MCMC sample of values for a specified model.

Returns the model updated with the MCMC sample and, in the case of block=0, the iter field incremented by 1.

m : model specification.

block : block for which to sample values (default: all blocks).

MCPhylo.unlist — Function

unlist(m::Model, block::Integer=0, transform::Bool=false)

unlist(m::Model, monitoronly::Bool)

unlist(m::Model, nodekeys::Vector{Symbol}, transform::Bool=false)

Convert (unlist) sets of logical and/or stochastic node values to vectors.

Returns vectors of concatenated node values.

m : model containing nodes to be unlisted or relisted.

block : sampling block of nodes to be listed (default: all blocks).

nodekeys : node(s) to be listed.

transform : whether to apply a link transformation in the conversion.

MCPhylo.relist — Function

relist(m::Model, x::AbstractArray{T}, block::Integer=0,
        transform::Bool=false) where {T<:Real}

relist(m::Model, x::AbstractArray{T}, block::Integer=0,
        transform::Bool=false) where {T<:GeneralNode}

relist(m::Model, x::AbstractArray{T},
        nodekeys::Vector{Symbol}, transform::Bool=false) where {T<:Any}

relist(m::Model, x::N, nodekeys::Vector{Symbol}, transform::Bool=false) where N<:GeneralNode

Reverse of unlist; ie. Converts vectors to sets of logical and/or stochastic node values. Same inputs and return values as unlist.

MCPhylo.relist! — Function

relist!(m::Model, x::AbstractArray{T}, block::Integer=0,
          transform::Bool=false) where {T<:Any}

relist!(m::Model, x::AbstractArray{T}, nodekey::Symbol,
              transform::Bool=false) where {T<:Real}

Reverse of unlist; ie. Converts vectors to sets of logical and/or stochastic node values. Same inputs as unlist.

Returns m, with values copied to the nodes.

Missing docstring.

Missing docstring for Main.MCPhylo.update!. Check Documenter's build log for details.