Utils

Functions found in all files of Utils folder, ordered according to file of origin.

utils

MCPhylo.assign_mcmc_work — Method

assignmcmcwork(f::Function, lsts::AbstractArray )::Tuple{Vector{Tuple{Chains, Model, ModelState}}, Array{Float64, 2}, Vector{AbstractString}}

–- INTERNAL –- Starts the MCMC chain generation (on multiple workers if possible) and also starts parallel ASDSF - if possible and requested by the user.

MCPhylo.lcp — Method

lcp(str1::T, str2::T)::T where T <: AbstractString

Get the longest common prefix.

MCPhylo.mcmc_or_convergence — Method

mcmcorconvergence(args::AbstractArray) ::Union{Tuple{Chains, Model, ModelState}}, Tuple{Vector{Vector{Float64}}, ConvergenceStorage}

–- INTERNAL –- Used in pmap call in assignmcmcwork to correctly dispatch arguments to either calculate convergence statistics or mcmc chains.