Tree Functionality

MCPhylo offers several tree functionalities. The underlying type is that of a GeneralNode. It provides the functionality necessary to do MCMC operations with tree structures.

Node Type

Main.MCPhylo.GeneralNode — Type

Node

This data type holds the basic Node structure. The type T is used to specify the type of the data stored in the node.

If nchild is 0 the Node is a leaf node.

If root is False the Node is a child of another node.

inc_length specifies the length of the incomming branch.

binary specifies the path from the root to the Node. 1 and 0 represent left and right turns respectively.

Main.MCPhylo.Node — Method

function Node()::Node

This function will initialize an empty node.

Converter

Main.MCPhylo.from_df — Function

from_df(df::DataFrame)::Node

This function takes a DataFrame and turns it into a tree. It assumes a rooted binary tree is stored in the matrix. No checks are performed.

Returns the root node of the tree.

df : Dataframe used to create a tree.

Main.MCPhylo.newick — Method

newick(node::Node)::String

Creates a newick representation of the tree.

Returns a properly formatted newick String.

node : root node of tree used to create the newick string.

Main.MCPhylo.to_covariance — Function

to_covariance(tree::N, blv::Array{T})::Array{T,2} where {N<:GeneralNode,T<: Real}

Calcualte the variance-covariance matrix from tree. An entry (i,j) of the matrix is defined as the length of the path connecting the latest common ancestor of i and j with the root of the tree.

Returns an Array of Real numbers.

tree : Node in tree of interest.

blv : branchlength vector of tree.

Main.MCPhylo.to_covariance_ultra — Function

to_covariance_ultra(tree::Node)::Array{T,2} where T<: Real

Get the covariance matrix of the ultrametric version of tree with height 1.

Returns an Array of Real numbers.

tree : root of tree used to perform calculation.

Main.MCPhylo.to_df — Function

to_df(root::Node)::DataFrame

This function returns a matrix representation of the tree structure. The matirx is returned as a DataFrame so that the names of the columns are the names of the tips in the tree. The entry df[i,j] is the length of the edge connecting node i with node j.

Returns a Dataframe.

root : root of tree used to create matrix represenation.

Main.MCPhylo.to_distance_matrix — Function

to_distance_matrix(tree::T)::Array{Float64,2} where T <:GeneralNode

Calculate the distance matrix over the set of leaves.

Returns an Array of Floats.

tree : root node of tree used to perform caclulcation.

Tree Basics

This file is a bit larger, and so will be broken into (hopefully coherent) subgroups, listed here:

Tree Editing

Return Functions

Metavariable Functions

Variable Functions

Tree Editing

Main.MCPhylo.add_child! — Function

add_child!(mother_node::Node, child::Node, child_position::Union{Int64, Missing}=missing)

This function adds a child to the mother node. The arity of the mother node is increased by 1 and the root status of the child is set to False.

mother_node : Node to add a child to.

child : Node to add to mother_node.children.

child_position : index at which to add the new child node; optional.

Main.MCPhylo.delete_node! — Function

delete_node!(node::Node)::Nothing

This functions deletes node from a tree and assigns all its children to its mother node.

node : Node to be deleted.

Main.MCPhylo.insert_node! — Function

insert_node!(mother::Node, children::Vector{T})::T where T<:AbstractNode

This function inserts a node into a tree after a mother node and gains a subset of the mother's children as its children.

Returns the inserted node.

mother : Node under which to add the newly-inserted node.

children : Children of node referenced by "mother" to reassign as children of the newly-inserted node.

Main.MCPhylo.remove_child! — Function

remove_child!(mother_node::Node, left::Bool)::Node

This function removes a child from the list of nodes which are daughters of this node. An input of "True" removes the left child, while "False" removes the right child.

Returns the removed node.

mother_node : Node from which to remove a child.

left : boolean value determining which child of mother_node to remove.

remove_child!(mother_node::Node, child::Node)::Node

This function removes a child from the list of nodes which are daughters of this node.

The removed node is returned.

mother_node : Node from which to remove a child.

child : specific Node to remove.

Main.MCPhylo.create_tree_from_leaves — Function

create_tree_from_leaves(leaf_nodes::Vector{T})::Node

This function creates a random binary tree from a list of leaf nodes.

The root node as access point for the tree is returned.

leaf_nodes : Names of nodes to create.

node_size : used to initialize node.data

cu : ??

Return Functions

Main.MCPhylo.get_mother — Function

function get_mother(node::T)::T  where T<:GeneralNode

This function gets the mother of node. It does so by looking for the respective binary representation of the mother node.

Main.MCPhylo.get_sister — Function

get_sister(node::T)::T  where T<:GeneralNode

This function gets the sister of node. It does so by looking for the respective binary representation of the sister.

Main.MCPhylo.node_height — Function

node_height(root::T, mv::Float64)::Float64  where T<:GeneralNode

Calculate the height of a node.

Main.MCPhylo.random_node — Function

function random_node(root::T)::T  where T<:GeneralNode

This function returns a random node from the tree.

Metavariable Functions

Main.MCPhylo.tree_length — Method

tree_length(root::T)::Float64  where T<:GeneralNode

This function calculates the tree length.

Main.MCPhylo.tree_height — Function

tree_height(root::T)::Float64  where T<:GeneralNode

This function calculates the tree height.

Main.MCPhylo.get_branchlength_vector — Function

get_branchlength_vector(root::N)::Vector{T}  where {N <:GeneralNode, T<:Real}

Get the vector of branch lengths of the tree.

get_branchlength_vector(t::TreeStochastic)

Get the vector of branch lengths of the tree.

get_branchlength_vector(root::N, out_vec::Vector{T}) where {N<:GeneralNode, T<:Real}

Do post order traversal to retrieve a vector of branch lengths.

get_branchlength_vector(post_order::Vector{Node})::Vector{Float64}

Return a vector of branch lengths.

post_order : Vector of Nodes of a tree.

Main.MCPhylo.get_sum_seperate_length! — Function

get_sum_seperate_length!(root::T)::Vector{Float64}  where T<:GeneralNode

This function gets the sum of the branch lengths of the internal branches and the branches leading to the leave nodes.

get_sum_seperate_length!(post_order::Vector{T})::Vector{Float64}  where T<:GeneralNode

This function gets the sum of the branch lengths of the internal branches and the branches leading to the leave nodes.

Main.MCPhylo.check_binary — Function

check_binary(root::Node)::Bool

checks to see if given tree is binary; returns true if properly formatted and false otherwise

Main.MCPhylo.path_length — Function

path_length(ancestor::T, descendant::T)::Float64  where T<:GeneralNode

Note: The function assumes there is an ancestral relationship between the two nodes.

This function calculates the length of the path separating the ancestor from the offspring node. The function follows the path specified through the binary description of the node.

Variable Functions

Main.MCPhylo.force_ultrametric! — Function

force_ultrametric!(root::T) where T<:GeneralNode

Force an ultrametric version of the tree.

Main.MCPhylo.number_nodes! — Function

number_nodes!(root::T)::Nothing  where T<:GeneralNode

This function assigns a unique, sequential number to each node. Leaves are numbered first in alphabetical order.

Main.MCPhylo.set_branchlength_vector! — Function

set_branchlength_vector!(t::TreeStochastic, blenvec::Array{T}) where T <: Real

Get the vector of branch lengths of the tree.

set_branchlength_vector!(t::TreeStochastic, blenvec::ArrayStochastic)

Get the vector of branch lengths of the tree.

set_branchlength_vector!(root::N, blenvec::Array{T}) where {N<:GeneralNode, T<:Real}

This function sets the branch lengths of a tree to the values specified in blenvec.

Main.MCPhylo.set_binary! — Function

set_binary!(root::Node)

Assign a binary representation to each node, which specifies the path from the root to this node via the binary representation of the node. A left turn is a 1 in binary and a right turn a 0.

Tree Clustering

Main.MCPhylo.neighbor_joining — Function

neighbor_joining(dm::Array{Float64,2}, Array{String,1})

This function returns a phylogenetic tree by using neighbor-joining based on a given distance matrix and an array of leaf names.

Returns a node of the resulting tree, from which it can be traversed.

dm : Matrix used to create Tree.

leaf_names : Array containing names of leaf nodes.

neighbor_joining(dm::Array{Float64,2})

This function returns a phylogenetic tree by using neighbor-joining based on a given distance matrix. Creates an array of nodes to be used as leaves.

Returns a node of the resulting tree, from which it can be traversed.

dm : Matrix from which to create tree.

Main.MCPhylo.upgma — Function

upgma(dm::Array{Float64,2}, Array{String,1})

This function returns a phylogenetic tree by using UPGMA based on a given distance matrix and an array of leaf names.

Returns a node of the resulting tree, from which it can be traversed.

dm : Matrix from which to create the tree.

leaf_names : array of strings containing names of leaf nodes.

upgma(dm::Array{Float64,2})

This function returns a phylogenetic tree by using UPGMA based on a given distance matrix. Creates an array of nodes to be used as leaves.

Returns a node of the resulting tree, from which it can be traversed.

dm : Matrix from which to create the tree.

Tree Consensus

Main.MCPhylo.check_node! — Method

check_node!(
    ref_node::T, leaves::Vector{T},
    leaf_ranks_reverse::Dict{String, Int64},
    xleft_dict::Dict{T, Tuple{T, Dict{Int64, T}}},
    xright_dict::Dict{T, Tuple{T, Dict{Int64, T}}},
    marked_nodes::Union{Dict{Int64, Bool, Nothing}}=nothing
)::Union{Nothing, Tuple{T, T, T, Dict{Int64, T}, Dict{Int64, T}}} where T<:AbstractNode

–- INTERNAL –- Helper function that handles major chunk of code that onewaycompatible and merge_tree share

Main.MCPhylo.count_cluster_occurences — Method

count_cluster_occurences(bit_vectors::BitVector)
    ::PriorityQueue{BitVector, Int64}

–- INTERNAL –- Helper function for greedyconsensustree that counts the occurrences of each bit vector representing a cluster. Returns a priority queue.

Main.MCPhylo.depth_dicts — Method

depth_dicts(leaves::Vector{T})
    ::Tuple{Dict{T, Tuple{T, Dict{Int64, T}}}, Dict{T, Tuple{T, Dict{Int64, T}}}} where T<:AbstractNode

–- INTERNAL –- Helper function for onewaycompatible and merge_trees!. Creates a dictionary based on a vector of leaf nodes, and stores the depth of each node, as well as the left and right path leading to it. Based on section 6.1 of the paper.

Main.MCPhylo.find_common_clusters — Method

find_common_clusters(ref_tree, tree:T)
    ::Dict{Int64, Tuple{Bool, Union{Float64, Missing}}}

–- INTERNAL –- Use Day's algorithm to create a dictionary, that tells us for each node of the second input tree, if its corresponding cluster is a common cluster of the trees.

Returns dictionary.

ref_tree : Tree used to create dictionary.

tree : Tree used to create dictionary.

Main.MCPhylo.get_cluster_start_indeces — Method

get_cluster_start_indeces(ref_nodes::T, tree::T)
    ::Dict{T, Int64} where T<:AbstractNode

–- INTERNAL –- Helper function to obtain the cluster start indeces for a tree (tree), based on the nodes of another tree (ref_nodes).

Main.MCPhylo.get_leaf_ranks — Method

function get_leaf_ranks(nodes::Vector{T})
    ::Dict{String, Int64} where T<:AbstractNode

–- INTERNAL –- Enumerate the leaf nodes in a tree. Returns a dictionary of this mapping.

Main.MCPhylo.greedy_consensus_tree — Method

greedy_consensus_tree(trees::Vector{T})::T where T<:AbstractNode

Construct the greedy consensus tree from a set of trees that share the same leafset. Returns the root node of the greedy consensus tree, from which it can be traversed. This algorithm is based on section 5 and 6.1 of:

Jesper Jansson, Chuanqi Shen, and Wing-Kin Sung. 2016. Improved algorithms for constructing consensustrees. J. ACM 63, 3, Article 28 (June 2016), 24 pages https://dl.acm.org/doi/pdf/10.1145/2925985

Main.MCPhylo.loose_consensus_tree — Method

loose_consensus_tree(trees::Vector{T})::T where T<:AbstractNode

Construct the loose consensus tree from a set of trees that share the same leafset. I.e. a tree with all the clusters that appear in at least one tree and are compatible with all trees. Returns the root node of the loose consensus tree, from which it can be traversed. This algorithm is based on section 4 and 6.1 of:

Jesper Jansson, Chuanqi Shen, and Wing-Kin Sung. 2016. Improved algorithms for constructing consensustrees. J. ACM 63, 3, Article 28 (June 2016), 24 pages https://dl.acm.org/doi/pdf/10.1145/2925985

Main.MCPhylo.majority_consensus_tree — Method

majority_consensus_tree(trees::Vector{T}, percentage::Float64=0.5)::T where T<:AbstractNode

Construct the majority rule consensus tree from a set of trees that share the same leafset. By default the output tree includes clusters that occur in over 50% of the trees. This can be customized when calling the function. The function returns the root node of the majority consensus tree, from which it can be traversed. The algorithm is based on section 3 and 6.1 of:

Jesper Jansson, Chuanqi Shen, and Wing-Kin Sung. 2016. Improved algorithms for constructing consensustrees. J. ACM 63, 3, Article 28 (June 2016), 24 pages https://dl.acm.org/doi/pdf/10.1145/2925985

Main.MCPhylo.max_leaf_rank — Method

max_leaf_rank(leaf_ranks::Dict{String, Int64}, node::T)
    ::Int64 where T<:AbstractNode

–- INTERNAL –- Recursive helper function to find the highest ranked leaf descendant of a node

Main.MCPhylo.merge_trees! — Method

merge_trees!(ref_tree::T, tree::T)::Vector{T}} where T<:AbstractNode

–- INTERNAL –- Merge two compatible trees, i.e. inserts all cluster of the first tree, which aren't already in the second tree, into the second tree

Returns a Tuple containing the root node of the merged tree, as well as a vector of nodes that were added.

ref_tree : tree used to determine common clusters.

tree : tree used to determine common clusters.

Main.MCPhylo.min_leaf_rank — Method

min_leaf_rank(leaf_ranks::Dict{String, Int64}, node::T)
    ::Int64 where T<:AbstractNode

–– INTERNAL –- Recursive helper function to find the lowest ranked leaf descendant of a node

Main.MCPhylo.one_way_compatible — Method

one_way_compatible(ref_tree::T, tree::T)::T where T<:AbstractNode

–- INTERNAL –- Takes two trees and returns a copy of the first one, where all the clusters that are not compatible with the second tree are removed.

Returns root node of edited tree.

ref_tree : tree used to determine clusters.

tree : tree used to determine clusters.

Main.MCPhylo.order_tree! — Method

order_tree!(root::T, cluster_start_indeces::Dict{T, Int64}, leaves=Vector{T}())
    ::Vector{T} where T<:AbstractNode

–- INTERNAL –- Helper function to order a tree based on cluster indeces and return the leaves of the ordered tree

Main.MCPhylo.set_node_stats! — Method

set_node_stats!(tree::T, trees::Vector{T}, majority::Bool)
    ::Nothing where T<:AbstractNode

–- INTERNAL –- Helper function for the construction of a consensus tree. Calculates the inc_lengths and statistics of the nodes in the consensus tree. If dealing with a (in-progress) majority consensus tree, this function will also delete its non-majority clusters (handled by the 'majority' boolean).

Main.MCPhylo.x_left — Method

x_left(node::T)::Tuple{T,Vector{T}} where T<:AbstractNode

–- INTERNAL –- Helper function to find ancestor of a leaf that has said leaf as leftmost descendant. Also returns the path from the leaf to the mother of that node.

Main.MCPhylo.x_right — Method

x_right(node::T)::Tuple{T,Vector{T}} where T<:AbstractNode

–- INTERNAL –- Helper function to find ancestor of a leaf that has said leaf as rightmost descendant. Also returns the path from the leaf to the mother of that node.

Tree Distance

Main.MCPhylo.BHV_bounds — Method

BHV_bounds(tree1::T, tree2::T)::Tuple{Float64, Float64} where T <:GeneralNode

This function calculates the lower and upper bounds of the geodesic in the Billera-Holmes-Vogtman space.

Returns tuple of floats.

Main.MCPhylo.RF — Method

RF(tree1::T, tree2::T)::Int64 where T <:GeneralNode

Calculate the Robinson-Foulds distance between the two trees. In its current form the function assumes the trees have identical leave sets.

Returns result of algorithm as integer.

tree1 : tree used to determine RF distance.

tree2 : tree used to determine RF distance.

Main.MCPhylo.get_bipartitions — Method

get_bipartitions(tree::T)::Vector{Tuple} where T <:GeneralNode

Get a vector of all bipartions of tree.

Returns a vector containing Tuples of sets representing the bipartions.

Tree Ladderizing

Main.MCPhylo.ladderize_tree! — Method

ladderize_tree!(root::T, ascending::Bool=true)::Nothing where T<:AbstractNode)

This function ladderizes a tree inplace, i.e. sorts the nodes on all levels by the count of their descendants.

root : root Node of tree.

ascending : Boolean, determines whether to sort in ascending (true) or descending (false) order.

Main.MCPhylo.ladderize_tree — Method

ladderize_tree(root::T, ascending::Bool=true)::T where T<:AbstractNode)

This function returns a ladderized copy of a tree, i.e. a copy with all the nodes on all levels sorted by the count of their descendants.

root : root Node of tree.

ascending : Boolean, determines whether to sort in ascending (true) or descending (false) order.

Tree Legacy

Main.MCPhylo.get_branchlength_vector — Method

get_branchlength_vector(post_order::Vector{Node})::Vector{Float64}

Return a vector of branch lengths.

post_order : Vector of Nodes of a tree.

Tree moves

Main.MCPhylo.NNI! — Function

NNI(root::T, target::T, lor::Bool)::Int64   where T<:GeneralNode

This function does a nearest neighbour interchange (NNI) move on the tree specified by root. The parameter target specifies the node which performs the interchange move using the left or right child of the target node. If the left child should be used lor=true.

The function returns 1 if the move was successful and 0 else.

root : root node of tree on which to perform the NNI.

target : specific node of tree to interchange.

lor : Bool; "true" uses the left child of target, "false," the right child.

NNI!(root::T, target::Int64)::Int64  where T<:GeneralNode

This function does a nearest neighbour interchange (NNI) move on the tree specified by root. The target is identified by the number of the target node.

The function returns 1 if the move was successful and 0 else.

root : root node of tree on which to perform the NNI.

target : specific node of tree to interchange.

NNI!(root::T, target::Int64)::Int64  where T<:GeneralNode

This function does a nearest neighbour interchange (NNI) move on the tree specified by root. The target is identified by the number of the target node. The function returns 1 if the move was successfull and 0 else.

NNI!(root::T)::Int64  where T<:GeneralNode

This function does a nearest neighbour interchange (NNI) move on the tree specified by root. The target is identified by the number of the target node.

The function returns 1 if the move was successful and 0 else.

root : root node of tree on which to perform the NNI.

Main.MCPhylo.NNI — Function

NNI(root::T)::T  where T<:GeneralNode

This function does a nearest neighbour interchange (NNI) move on the tree specified by root.

Returns a mutated copy while leaving the original tree intact.

root : root node of tree on which to perform the NNI.

Main.MCPhylo.change_edge_length! — Function

change_edge_length!(root::T) where T <:GeneralNode

Pick a random node and increase or decrease its length randomly.

root : root node of tree.

Main.MCPhylo.move! — Function

move!(node1::Node, node2::Node, proportion::Float64)

Change the incoming length of node1 and node2 while keeping their combined length constant.

node1 : Node whose inclength will be modified; this node's inclength will be the total inc_length of both nodes, times proportion.

node2 : Node whose inclength will be modified; this node's inclength will be the remainder of total - the new inc_length value of node1.

proportion : Float64, determines proportion of the inc_length of both nodes assigned to node1.

Main.MCPhylo.perform_spr — Function

perform_spr(root::Node)

performs SPR on binary tree.

Returns root of tree post-SPR.

root : Node of tree on which to perform SPR.

Main.MCPhylo.randomize! — Function

randomize!(root::Node, num::Int64=100)::nothing

This function randomizes the tree topology by performing a number of nearest neighbour interchange (NNI) moves. The number of NNI moves is specified in the parameter num.

root : root node of tree to be edited.

num : amount of NNI moves to perform.

Main.MCPhylo.risky_SPR! — Function

    risky_SPR!(root::Node)::AbstractNode

Performs SPR on tree in place.

Returns reference to root of altered tree. Does not check for correct formatting of tree.

root : root node of tree.

Main.MCPhylo.risky_SPR — Function

    risky_SPR(root::Node)::AbstractNode
Performs SPR on tree in place. Takes reference to root of tree
Returns copy of root of altered tree. Does not check for correct formatting of tree.

Main.MCPhylo.slide — Function

slide(root::T)::T where T<:GeneralNode

This functin performs a slide move on an intermediate node. The node is moved upwards or downwards on the path specified by its mother and one of its daughters.

Returns root Node of new tree.

root : root Node of tree.

Main.MCPhylo.slide! — Function

slide!(root::T) where T<:GeneralNode

This functin performs a slide move on an intermediate node. The node is moved upwards or downwards on the path specified by its mother and one of its daughters.

root : root Node of tree.

Main.MCPhylo.swing! — Function

swing!(root::T) where T<:GeneralNode

This function performs a swing node. A random non-leave node is selected and moved along the path specified by its two children.

root : root Node of tree.

Main.MCPhylo.swing — Function

swing(root::T)::T where T<:GeneralNode

This function performs a swing node. A random non-leave node is selected and moved along the path specified by its two children.

Returns root Node of new tree.

root : root Node of tree.

Tree Pruning

Main.MCPhylo.prune_tree! — Function

prune_tree!(root::T, node_names::Vector{String})::Nothing where T<:AbstractNode

In-place version of prune_tree.

root : root Node of tree to prune.

node_names : vector of strings, used to specify nodes to remove.

prune_tree!(root::T, node_names::Vector{T})::Nothing where T<:AbstractNode

In-place version of prune_tree.

root : root node of tree to prune.

node_names: vector of Node objects to be removed from tree.

Main.MCPhylo.prune_tree — Function

prune_tree(root::T, node_names::Vector{String})::T where T<:AbstractNode

This function removes specific nodes, including their descendants, from a tree.

root : root Node of tree to prune.

node_names : vector of strings, used to specify nodes to remove.

Tree Search

Main.MCPhylo.find_binary — Function

find_binary(root::T, bin::String)::T where T<:GeneralNode

Find a node by its binary representation. The function assumes that the node is present in the tree.

Do not use this function if you are unsure whether the node is in the tree at all.

Returns a reference to the desired Node.

root : root Node of tree to search.

bin : binary representation of desired Node as a String.

Main.MCPhylo.find_num — Function

find_num(root::T, num::Int64)  where T<:GeneralNode

Find a node by its number. The function assumes that the node is present in the tree.

Do not use this function if you are unsure whether the node is in the tree at all.

Returns reference to Node.

root : root Node of tree to be searched.

num : number of desired Node.

find_num(root::T, num::Int64, rn::Vector{T})::Bool  where T<:GeneralNode

Do a post order traversal to find the node corresponding to the num.

Returns true if node is found, false otherwise. Desired Node is pushed to rn.

root : root Node of tree to be searched.

num : number of desired Node.

rn : Vector of Nodes; desired Node is pushed to this vector when found.

Main.MCPhylo.find_root — Function

find_root(node::Node)::Node

Finds the root of tree indicated by Node.

Returns reference to root Node of the tree.

node : Node in Tree of interest.

Tree Traversal

Main.MCPhylo.get_leaves — Function

get_leaves(root::T, traversal::Vector{T})::Vector{T} where T<:GeneralNode

This function returns leaf nodes of a tree. It is assumed that root is the root of the tree. Thus, if root is not the root, the subtree defined by the root root is used.

Returns a vector of leaf Nodes.

root : root Node of tree to traverse.

traversal : vector of Nodes; leaf Nodes, once found, are pushed to this vector.

get_leaves(root::T)::Vector{T} where T<:GeneralNode

This function returns the leaves of a tree. Only the root node needs to be supplied.

Returns vector of leaf Nodes.

root : root Node of tree.

Main.MCPhylo.level_order — Function

level_order(node::T)::Array{T} where T<:GeneralNode

This function does level order traversal. Only the root node needs to be supplied.

Returns Array of Nodes.

node : root Node of tree.

Main.MCPhylo.level_traverse — Function

level_traverse(node::T, level::Int64, stack::Array{T})::Bool where T <:GeneralNode

This function traverses a level of the tree specified through node. The level is specified via the level argument and the nodes visited are stored in the stack. This function is intended as the internal worker for the level_order function.

Main.MCPhylo.post_order — Function

post_order(root::T, traversal::Vector{T})::Vector{T} where T<:GeneralNode

This function performs a post order traversal through the tree. It is assumed that root is the root of the tree. Thus, if root is not the root, the subtree defined by the root root is used for the post order traversal.

Returns vector of Nodes.

root : root Node of tree.

traversal : vector of Nodes; nodes are pushed to this vector as the tree is traversed.

post_order(root::T)::Vector{T} where T<:GeneralNode

This function does post order traversal. Only the root node needs to be supplied.

Returns vector of Nodes.

root : root Node of tree.

Main.MCPhylo.pre_order — Function

pre_order(root::T, traversal::Vector{T})::Vector{T} where T<:GeneralNode

This function performs a pre order traversal through the tree. It is assumed that root is the root of the tree. Thus, if root is not the root, the subtree defined by the root root is used for the pre order traversal.

Returns vector of Nodes.

root : root Node of tree.

traversal : vector of Nodes; nodes are pushed to this vector as the tree is traversed.

pre_order(root::T)::Vector{T} where T<:GeneralNode

This function does pre order traversal. Only the root node needs to be supplied.

Returns vector of Nodes.

root : root Node of tree.