List of all functions (minus private ones) in the alphabetical order, capital letters first

Main.MCPhylo.BHV_bounds — Method

BHV_bounds(tree1::T, tree2::T)::Tuple{Float64, Float64} where T <:GeneralNode

This function calculates the lower and upper bounds of the geodesic in the Billera-Holmes-Vogtman space.

Main.MCPhylo.NNI! — Method

NNI!(root::T, target::Int64)::Int64  where T<:GeneralNode

This function does a nearest neighbour interchange (NNI) move on the tree specified by root. The target is identified by the number of the target node. The function returns 1 if the move was successfull and 0 else.

Main.MCPhylo.NNI! — Method

NNI(root::T, target::T, lor::Bool)::Int64   where T<:GeneralNode

This function does a nearest neighbour interchange (NNI) move on the tree specified by root. The parameter target specifies the node which performs the interchange move using the left or right child of the target node. If the left child should be used lor=true. The function returns 1 if the move was successfull and 0 else.

Main.MCPhylo.NNI! — Method

NNI!(root::T)::Int64  where T<:GeneralNode

This function does a nearest neighbour interchange (NNI) move on the tree specified by root. The target is identified by the number of the target node. The function returns 1 if the move was successfull and 0 else.

Main.MCPhylo.NNI — Method

NNI(root::T)::T  where T<:GeneralNode

This function does a nearest neighbour interchange (NNI) move on the tree specified by root and returns a mutated copy while leaving the original tree intact.

Main.MCPhylo.RF — Method

RF(tree1::T, tree2::T)::Int64 where T <:GeneralNode

Calculate the Robinson-Foulds distance between the two trees. In its current form the function assumes the trees have identical leave sets.

Main.MCPhylo.RWM — Method

RWM(params::ElementOrVector{Symbol}, moves::Array{Symbol}; args...)

Construct the RWM sampler for Trees. If you set moves to :all it will use all eligible moves to change the tree. These are currently: NNI, Slide, Swing, :EdgeLength

Main.MCPhylo.add_child! — Function

add_child!(mother_node::Node, child::Node, child_position::Union{Int64, Missing}=missing)

This function adds a child to the mother node. The arity of the mother node is increased by 1 and the root status of the child is set to False.

Main.MCPhylo.create_tree_from_leaves — Function

create_tree_from_leaves(leaf_nodes::Vector{T})::Node

This function creates a random binary tree from a list of leaf nodes. The root node as access point for the tree is returned.

Main.MCPhylo.delete_node! — Method

delete_node!(node::Node)::Nothing

This functions deletes node from a tree and assigns all its children to its mother node.

Main.MCPhylo.find_root — Method

find_root(node::Node)::Node

Takes a node in a tree and finds the root of that tree

Main.MCPhylo.get_branchlength_vector — Method

get_branchlength_vector(t::TreeStochastic)

Get the vector of branch lengths of the tree.

Main.MCPhylo.get_branchlength_vector — Method

get_branchlength_vector(root::N)::Vector{T}  where {N <:GeneralNode, T<:Real}

Get the vector of branch lengths of the tree.

Main.MCPhylo.get_branchlength_vector — Method

get_branchlength_vector(post_order::Vector{Node})::Vector{Float64}

Return a vector of branch lengths.

Main.MCPhylo.get_branchlength_vector — Method

get_branchlength_vector(root::N, out_vec::Vector{T}) where {N<:GeneralNode, T<:Real}

Do post order traversal to retrieve a vector of branch lengths.

Main.MCPhylo.get_leaves — Method

post_order(root::T, traversal::Vector{T})::Vector{T} where T<:GeneralNode

This function performs a post order traversal through the tree. It is assumed that root is the root of the tree. Thus, if root is not the root, the subtree defined by the root root is used for the post order traversal.

Main.MCPhylo.get_leaves — Method

post_order(root::T)::Vector{T} where T<:GeneralNode

This function does post order traversal. Only the root node needs to be supplied.

Main.MCPhylo.get_sister — Method

get_sister(node::T)::T  where T<:GeneralNode

This function gets the sister of node. It does so by looking for the respective binary representation of the sister.

Main.MCPhylo.insert_node! — Method

insert_node!(mother::Node, children::Vector{T})::T where T<:AbstractNode

This function inserts a node into a tree after a mother node and gains a subset of the mother's children as its children. Returns the inserted node.

Main.MCPhylo.ladderize_tree! — Method

ladderize_tree!(root::T, ascending::Bool=true)::Nothing where T<:AbstractNode)

This function ladderizes a tree inplace, i.e. sorts the nodes on all levels by the count of their descendants

Main.MCPhylo.ladderize_tree — Method

ladderize_tree(root::T, ascending::Bool=true)::T where T<:AbstractNode)

This function returns a ladderized copy of a tree, i.e. a copy with all the nodes on all levels sorted by the count of their descendants

Main.MCPhylo.level_order — Method

level_order(node::T)::Array{T} where T<:GeneralNode

This function does level order traversal. Only the root node needs to be supplied.

Main.MCPhylo.majority_consensus_tree — Method

majority_consensus_tree(trees::Vector{T}, percentage::Float64=0.5)::T where T<:AbstractNode

Construct the majority rule consensus tree from a set of trees. By default includes cluster that occur in over 50% of the trees.

Main.MCPhylo.neighbor_joining — Method

neighbor_joining(dm::Array{Float64,2}, Array{String,1})

This function returns a phylogenetic tree by using neighbor-joining based on a given distance matrix and an array of leaf names. Returns a node of the resulting tree, from which it can be traversed.

Main.MCPhylo.neighbor_joining — Method

neighbor_joining(dm::Array{Float64,2})

This function returns a phylogenetic tree by using neighbor-joining based on a given distance matrix. Creates an array of nodes to be used as leaves. Returns a node of the resulting tree, from which it can be traversed.

Main.MCPhylo.newick — Method

newick(root::T, newickstring::AbstractString) where T<:GeneralNode

Do the newick recursion. It is meant as the internal iterator function.

Main.MCPhylo.newick — Method

newick(node::Node)::String

Creates a newick represnetation of the tree.

Main.MCPhylo.path_length — Method

path_length(ancestor::T, descendant::T)::Float64  where T<:GeneralNode

Note: The function assumes there is an ancestral relationship between the two nodes.

This function calculates the length of the path separating the ancestor from the offspring node. The function follows the path specified through the binary description of the node.

Main.MCPhylo.post_order — Method

post_order(root::T, traversal::Vector{T})::Vector{T} where T<:GeneralNode

This function performs a post order traversal through the tree. It is assumed that root is the root of the tree. Thus, if root is not the root, the subtree defined by the root root is used for the post order traversal.

Main.MCPhylo.post_order — Method

post_order(root::T)::Vector{T} where T<:GeneralNode

This function does post order traversal. Only the root node needs to be supplied.

Main.MCPhylo.pre_order — Method

pre_order(root::T, traversal::Vector{T})::Vector{T} where T<:GeneralNode

This function performs a pre order traversal through the tree. It is assumed that root is the root of the tree. Thus, if root is not the root, the subtree defined by the root root is used for the pre order traversal.

Main.MCPhylo.pre_order — Method

pre_order(root::T)::Vector{T} where T<:GeneralNode

This function does pre order traversal. Only the root node needs to be supplied.

Main.MCPhylo.prune_tree! — Method

prune_tree!(root::T, node_names::Vector{String})::Nothing where T<:AbstractNode

In-place version of prune_tree

Main.MCPhylo.prune_tree! — Method

prune_tree!(root::T, node_names::Vector{T})::Nothing where T<:AbstractNode

In-place version of prune_tree

Main.MCPhylo.prune_tree — Method

prune_tree(root::T, node_names::Vector{String})::T where T<:AbstractNode

This function returns a copy of a tree with specific nodes including their descendants removed

Main.MCPhylo.randomize! — Method

randomize!(root::Node, num::Int64=100)::nothing

This function randomizes the tree topology by performing a number of nearest neighbour interchange (NNI) moves. The number of NNI moves is specified in the parameter num.

Main.MCPhylo.remove_child! — Method

remove_child!(mother_node::Node, left::Bool)::Node

This function removes a child from the list of nodes which are daughters of this node. The removed node is returned.

Main.MCPhylo.remove_child! — Method

remove_child!(mother_node::Node, child::Node)::Node

This function removes a child from the list of nodes which are daughters of this node. The removed node is returned.

Main.MCPhylo.set_branchlength_vector! — Method

set_branchlength_vector!(t::TreeStochastic, blenvec::Array{T}) where T <: Real

Get the vector of branch lengths of the tree.

Main.MCPhylo.set_branchlength_vector! — Method

set_branchlength_vector!(t::TreeStochastic, blenvec::Array{T}) where T <: Real

Get the vector of branch lengths of the tree.

Main.MCPhylo.set_branchlength_vector! — Method

set_branchlength_vector!(root::N, blenvec::Array{T}) where {N<:GeneralNode, T<:Real}

This function sets the branch lengths of a tree to the values specified in blenvec.

Main.MCPhylo.slide! — Method

slide!(root::T) where T<:GeneralNode

This functin performs a slide move on an intermediate node. The node is moved upwards or downwards on the path specified by its mother and one of its daughters.

Main.MCPhylo.slide — Method

slide(root::T)::T where T<:GeneralNode

This functin performs a slide move on an intermediate node. The node is moved upwards or downwards on the path specified by its mother and one of its daughters. The new tree is returned.

Main.MCPhylo.swing! — Method

swing!(root::T) where T<:GeneralNode

This function performs a swing node. A random non-leave node is selected and moved along the path specified by its two children.

Main.MCPhylo.swing — Method

swing(root::T)::T where T<:GeneralNode

This function performs a swing node. A random non-leave node is selected and moved along the path specified by its two children. The new tree is returned.

Main.MCPhylo.to_covariance — Method

to_covariance(tree::N, blv::Array{T})::Array{T,2} where {N<:GeneralNode,T<: Real}

Calcualte the variance-covariance matrix from tree. An entry (i,j) of the matrix is defined as the length of the path connecting the latest common ancestor of i and j with the root of the tree.

Main.MCPhylo.to_df — Method

to_df(root::Node)::DataFrame

This function returns a matrix representation of the tree structure. The matirx is returned as a DataFrame so that the names of the columns are the names of the tips in the tree. The entry df[i,j] is the length of the edge connecting node i with node j.

Main.MCPhylo.to_distance_matrix — Method

to_distance_matrix(tree::T)::Array{Float64,2} where T <:GeneralNode

Calculate the distance matrix over the set of leaves.

Main.MCPhylo.tree_height — Method

tree_height(root::T)::Float64  where T<:GeneralNode

This function calculates the tree height.

Main.MCPhylo.tree_length — Method

tree_length(root::T, tl::Float64)::Float64 where T<:GeneralNode

This function does the internal tree length recursion

Main.MCPhylo.tree_length — Method

tree_length(root::T)::Float64  where T<:GeneralNode

This function calculates the tree length.

Main.MCPhylo.upgma — Method

upgma(dm::Array{Float64,2}, Array{String,1})

This function returns a phylogenetic tree by using UPGMA based on a given distance matrix and an array of leaf names. Returns a node of the resulting tree, from which it can be traversed.

Main.MCPhylo.upgma — Method

upgma(dm::Array{Float64,2})

This function returns a phylogenetic tree by using UPGMA based on a given distance matrix. Creates an array of nodes to be used as leaves. Returns a node of the resulting tree, from which it can be traversed.