Parser

Parser functions create DataFrame and/or Tree representations from data stored as various filetypes.

CSV

Main.MCPhylo.ParseCSV — Function

ParseCSV(filename::String, header::Bool=true)

This function parses a CSV file containing input for the MCMC compuation. The file should follow the conventions used for MrBayes. For example:

Swedish_0,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,?,0,0,?,0,0
Welsh_N_0,0,0,0,0,0,0,?,0,0,0,0,?,?,0,0,?,0,0,0,1,?,?,0
Sardinian_N_0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,1,0,0,?,0,0,0,0,0

Set input for "header" to false if no header is present in the file.

Returns ntax, nchar as Integer values, gap as a String; df is returned as a DataFrame, and stores language names and data.

filename : Name of CSV file to parse.

header : Boolean value. "true" denotes there is a header to be skipped; input "false" if the file does not contain a header.

Main.MCPhylo.create_csvdf — Function

function create_csvdf(filecontnt::Array{String}, separator::AbstractString=",")::DataFrame

This function parses a CSV file and returns its content as a DataFrame.

NOTE: May later on be replaced by the respective DataFrames function.

Newick

Main.MCPhylo.ParseNewick — Method

ParseNewick(filename::String)::Array{AbstractNode, 1}

This function takes a filename as a String, and returns an array of trees(represented as Node objects). The file should solely consist of newick tree representations, separated by line. The function checks for proper newick formatting, and will return an error if the file is incorrectly formatted.

Returns an Array of Nodes; each Node is the root of the tree represented by a newick string in the file.

filename : name of file containing newick strings to parse.

Nexus

Main.MCPhylo.ParseNexus — Method

ParseNexus(filename::String)

This function parses a NEXUS file which stores the input for the MCMC compuation. The file should follow the conventions used for MrBayes.

Returns ntax, nchar, gap, and missing_representation values; returns Dataframe storing language names and data.

filename : NEXUS file to be parsed.

Main.MCPhylo.create_nexusdf — Method

create_nexusdf(filecontent::Array{String})::DataFrame

This function creates a DataFrame of the actual data. Used in ParseNexus().

Returns DataFrame of language names and data derived from NEXUS file.

filecontent : Array of Strings; Strings are read from NEXUS file in ParseNexus().

Main.MCPhylo.extract_meta_info — Method

extract_meta_info(content::Array{String})

This function extracts some meta information from the content of the nexus file. It "eats-up" the stack.

Returns values derived from metadata (ntax, nchar, gap, missing_representation). Used in ParseNexus().

content : Array of Strings; Strings are read from NEXUS file in ParseNexus().

General Parser Functions

Main.MCPhylo.make_tree_with_data — Function

make_tree_with_data(filename::String, dialect::AbstractString="nexus",
						 gap::Union{Missing, AbstractString}=missing,
						 miss::Union{Missing,AbstractString}=missing,
						 header::Bool=false)

General parsing function; user specifies type of file to parse.

Returns root Node of tree derived from data, as well as a DataFrame.

filename : Name of file to parse.

dialect : Specifies filetype to parse; currently accepts EITHER "nexus" or "csv" as inputs.

gap : specifies gap symbol for a CSV file; if "dialect" == "csv", this input is required, can be ommitted otherwise.

miss : specifies missing symbol for a CSV file; if "dialect" == "csv", this input is required, can be ommitted otherwise.

header : Boolean required for parsing CSV file, specifies if file contains a header; input not required when parsing a NEXUS file.

binary : Boolean, specifies whether a binary or nonbinary tree should be created for the given file. Defaults to true/binary tree.

Main.MCPhylo.make_tree_with_data_cu — Function

make_tree_with_data_cu(filename::String, dialect::AbstractString="nexus",
						 gap::Union{Missing, AbstractString}=missing,
						 miss::Union{Missing,AbstractString}=missing,
						 header::Bool=false)

General parsing function; user specifies type of file to parse.

Returns root Node of tree derived from data, as well as a CuArray-stored DataFrame.

filename : Name of file to parse.

dialect : Specifies filetype to parse; currently accepts EITHER "nexus" or "csv" as inputs.

gap : specifies gap symbol for a CSV file; if "dialect" == "csv", this input is required, can be ommitted otherwise.

miss : specifies missing symbol for a CSV file; if "dialect" == "csv", this input is required, can be ommitted otherwise.

header : Boolean required for parsing CSV file, specifies if file contains a header; input not required when parsing a NEXUS file.