samplers

Functions found in all files of Parser folder, ordered according to file of origin.

rwm.jl

Main.MCPhylo.RWM — Method

RWM(params::ElementOrVector{Symbol}, moves::Array{Symbol}; args...)

Construct the RWM sampler for Trees. If you set moves to :all it will use all eligible moves to change the tree. These are currently: NNI, Slide, Swing, :EdgeLength